% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_semla.R
\name{AddSemlaToSeurat}
\alias{AddSemlaToSeurat}
\title{Update a Seurat object for compatibility with semla}
\usage{
AddSemlaToSeurat(
  object,
  image_type = c("tissue_lowres", "tissue_hires"),
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{Seurat} with Visium data}

\item{image_type}{One of "tissue_lowres" or "tissue_hires", specifying
what H&E image was loaded into the \code{Seurat} object}

\item{verbose}{Print messages}
}
\value{
A \code{Seurat} object compatible with semla
}
\description{
In order to make a \code{Seurat} object with \code{semla}, 
you can use \code{AddSemlaToSeurat} to add the required 
data to the \code{Seurat} object. This assumes that 
the \code{Seurat} object contains one or more "VisiumV1" 
object(s) in the \code{images} slot.
}
\details{
Note that you need to specify what H&E image was loaded, 
one of "tissue_lowres" or "tissue_hires". If this argument 
is incorrect, the tissue coordinates will be misplaced.

Visit the \href{https://ludvigla.github.io/semla/articles/getting_started.html}{getting started}
tutorial on our package website for more details on how to 
use this function.
}
\examples{

\dontrun{
library(semla)
library(SeuratData)

InstallData("stxBrain")

# Load example Seurat object
brain <- LoadData("stxBrain", type = "anterior1")

# Make Seurat object compatible with semla
brain_semla <- AddSemlaToSeurat(brain)
}

}
