% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_spatial_network_utils.R
\name{AdjustTissueCoordinates}
\alias{AdjustTissueCoordinates}
\title{Adjust tissue coordinates for digital unrolling}
\usage{
AdjustTissueCoordinates(full_graph, verbose = TRUE)
}
\arguments{
\item{full_graph}{A `tbl_graph` object generated with \code{\link{CutSpatialNetwork}}}

\item{verbose}{Print messages}
}
\value{
A `tibble` with "unrolled" tissue coordinates
}
\description{
This function takes a `tbl_graph` object as input generated
with \code{\link{CutSpatialNetwork}} and attempts to calculate
the unrolled tissue coordinates.
}
\details{
First, the end point of the graph
are identified. The outermost end point is assumed to be the starting
point and the point at the center of the roll is assumed to be the
end point. Then, the algorithm tries to find the geodesic between the
two end points, forming a band of nodes which represents the shortest
path between the starting point and the end point. The order of these
shortest represents the distances along the x axis in the new coordinate
system and the geodesics to all other nodes are used as distances on the
y axis of the in the new coordinate system. By using information about the
location of nodes relative to the position of the shortest path nodes, the
algorithm also tries to adjust the y axis to ensure non-negative values.

NB: The graph has to be fully connected. If multiple sub graphs are found,
only the largest graph will be kept. It is crucial that edges has been cut
properly with \code{\link{CutSpatialNetwork}}, otherwise the results will be
inaccurate. See the package website for examples.
}
