% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_distance_plot.R
\name{centroid_angles_plot}
\alias{centroid_angles_plot}
\title{Draw a centroid angle plot}
\usage{
centroid_angles_plot(nbreaks = 9, centroid_size = 8)
}
\arguments{
\item{nbreaks}{Number of intervals to cut the angles into}

\item{centroid_size}{Size of the centroid spot}
}
\value{
A \code{ggplot} object
}
\description{
Draw a centroid angle plot
}
\examples{
# Draw a plot
centroid_angles_plot(9)

}
