% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_coordinates.R
\name{CoordAndImageTransform}
\alias{CoordAndImageTransform}
\title{Apply transformation to paired image and coordinates}
\usage{
CoordAndImageTransform(im, xy_coords, angle = 0, xy_offset = NULL)
}
\arguments{
\item{im}{Image of class `magick-image`, `StoredSpatialImage`, `raster` or a
path to an external image file.}

\item{xy_coords}{A \code{matrix}, \code{data.frame} or \code{tibble} object with numeric x, y coordinates.}

\item{angle}{Numeric value specifying the degree of rotation. Use negative angles
for  counter-clockwise rotation.}

\item{xy_offset}{Optional point (x, y) specifying the translation.}
}
\value{
A list with two elements:
\itemize{
  \item{"im_transf": An object of class `magick-image` representing the transformed image}
  \item{"xy_transf": An object of class `tble` representing the transformed coordinates}
}
}
\description{
SRT data generated with Visium consists of an H&E image and a gene expression
matrix. The columns of the gene expression matrix correspond to spots that can
be mapped to the H&E image using a set of pixel coordinates.
This function takes an image and its corresponding set of spot coordinates as
input and applies the same transformation to the image and spot coordinates
simultaneously.
}
\seealso{
Other transforms: 
\code{\link{CoordMirror}()},
\code{\link{CoordTransform}()},
\code{\link{ImageTranslate}()}
}
\concept{transforms}
