% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_coordinates.R
\name{CoordMirror}
\alias{CoordMirror}
\title{Mirror coordinates}
\usage{
CoordMirror(xy_coords, mirror.x = FALSE, mirror.y = FALSE, center = NULL)
}
\arguments{
\item{xy_coords}{A \code{matrix}, \code{data.frame} or \code{tibble} object with numeric x, y coordinates.}

\item{mirror.x, mirror.y}{Logical specifying whether the coordinates should be reflected along
the x-axis, y-axis or both.}

\item{center}{Optional point (x, y) specifying the center of reflection.}
}
\value{
A \code{tbl} object with transformed coordinates
}
\description{
Mirror coordinates
}
\details{
The coordinate system for \code{xy_coords} should match the dimensions of the
image. In other words, the coordinates should map spots to the tissue section on H&E image.
A 3x3 transformation matrix is constructed by combining the following matrices:

\itemize{
\item{\eqn{T(-x, -y)}: Translate coordinates to origin, i.e. (0, 0) becomes the new center \cr\tabular{llc}{
   1 \tab 0 \tab \eqn{-center_{x}} \cr
   0 \tab 1 \tab \eqn{-center_{y}} \cr
   0 \tab 0 \tab 1 \cr
}


}
\item{\eqn{M_{x}}: Reflect coordinates along x-axis\tabular{lll}{
   -1 \tab 0 \tab 0 \cr
   0 \tab 1 \tab 0 \cr
   0 \tab 0 \tab 1 \cr
}


}
\item{\eqn{M_{y}}: Reflect coordinates along y-axis \cr\tabular{lll}{
   1 \tab 0 \tab 0 \cr
   0 \tab -1 \tab 0 \cr
   0 \tab 0 \tab 1 \cr
}


}
\item{\eqn{T(x, y)}: Translate coordinates back to center \cr\tabular{llc}{
   1 \tab 0 \tab \eqn{center_{x}} \cr
   0 \tab 1 \tab \eqn{center_{y}} \cr
   0 \tab 0 \tab 1 \cr
}


}
}

Then, these matrices are combined to form the final transformation matrix:

\eqn{T_{final} = T(x, y)*M_{y}*M_{x}*T(-x, -y)}

Which can be used to transform our input coordinates:

\eqn{xy_{out} = T_{final}*xy_{in}}
}
\examples{
library(ggplot2)

# Create a data.frame with x, y coordinates
xy <- data.frame(x = 1:20, y = 1:20)

# Reflect coordinates along x axis
xy_mx <- CoordMirror(xy, mirror.x = TRUE) |> setNames(nm = c("x", "y"))

# Reflect along both x, and y axes
xy_mxy <- CoordMirror(xy, mirror.x = TRUE, mirror.y = TRUE) |> setNames(nm = c("x", "y"))

# Combine all coordinates
xy_all <- do.call(rbind, list(cbind(xy, type = "original", ord = 1:20),
                    cbind(xy_mx, type = "mirror_x", ord = 1:20),
                    cbind(xy_mxy, type = "mirror_x_and_y", ord = 1:20)))
xy_all$type <- factor(xy_all$type, levels = c("original", "mirror_x", "mirror_x_and_y"))

# Now we can sse the effects of mirroring
# Mirror x flips the coordinates along the x axis while mirror_x and mirror_y
# effectively inverts the coordinates, thus changing the order of the points
ggplot(xy_all, aes(x, y)) +
  geom_point(color = "steelblue", size = 7, alpha = 0.5) +
  geom_text(aes(label = ord)) +
  facet_grid(~type) +
  geom_vline(xintercept = 10.5, linetype = "dashed", color = "red") +
  geom_hline(yintercept = 10.5, linetype = "dashed", color = "green")

}
\seealso{
Other transforms: 
\code{\link{CoordAndImageTransform}()},
\code{\link{CoordTransform}()},
\code{\link{ImageTranslate}()}
}
\concept{transforms}
