% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_autocorrelation.R
\name{CorSpatialGenes}
\alias{CorSpatialGenes}
\title{Find features with high spatial autocorrelation}
\usage{
CorSpatialGenes(
  x,
  spatnet,
  across_all = FALSE,
  nCores = detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{x}{A matrix-like object with features in columns and spots in rows}

\item{spatnet}{A list of spatial networks created with \code{GetSpatialNetwork}. The spots in these
networks should match the spots in \code{x}.}

\item{across_all}{Should the autocorrelation scores be calculated across all samples?}

\item{verbose}{Print messages}
}
\value{
Either a list of tibbles or a tibble with gene names and correlation scores for each gene in \code{x}
}
\description{
This function can be used to find genes with high spatial autocorrelation in SRT data.
A more detailed description of the algorithm is outlined in the Details section below.
}
\section{Spatial autocorrelation}{

Spatial autocorrelation of genes is the term used to describe the presence of systematic spatial
variation in the expression of a gene. Positive spatial autocorrelation of a gene is the tendency
for regions that are close together in space to have similar expression levels.

A simple example is when you have an anatomical structure or a tissue type that spans
across multple neighboring spots in an SRT experiment, for example a gland, an immune
infiltrate or a region of the brain. Inside such structures, you might find that the
expression levels of certain genes are highly similar and hence these genes have a
positive spatial autocorrelation.

The method provided in \code{STUtility2} is relatively simple and fast. For each gene and spot,
the expression is averaged across all neighboring spots (typically the 6 closest neighbors)
to produce a lag expression vector. Since this vector represents the average of the surrounding
spots, we can use it to test if the expression in those spots is similar to the center spot.
One simple strategy is to calculate the pearson correlation between a genes' lag vector and
the original expression vector which typically captures the spatial autocorrelation well.
}

\section{Method steps}{

\itemize{
\item{Load a matrix with features in rows and spots in columns: \eqn{X_{expr}}}
\item{Convert the corresponding spatial network to wide format and construct a nearest
neighbor matrix \eqn{N_{neighbors}} in which neighboring spots have a value of 1
and the remaining spots have a value of 0
}
\item{\eqn{N_{neighbors}} is then multiplied with the \eqn{X_{expr}} to
calculate a lag vector for each gene: \cr \cr
\eqn{X_{lagexpr} = (N_{neighbors}*X_{expr})/n_{neighbors}} \cr \cr
where \eqn{n_{neighbors}} is the number of neighbors for each spot.
}
\item{The spatial autocorrelation score for a genes is the 'pearson' correlation of the
lag vector and the initial expression vector: \cr \cr
\eqn{spatcor_{gene} = cor(X_{lagexpr}[gene, ], X_{expr}[gene, ])}
}
}
}

\examples{
\dontrun{
# TODO
}

}
\seealso{
Other network-methods: 
\code{\link{GetSpatialNetwork}()}
}
\concept{network-methods}
