% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_unroll.R,
%   R/prepare_for_digital_unroll.R
\name{CutSpatialNetwork}
\alias{CutSpatialNetwork}
\alias{export_graph}
\title{Manually unroll tissues}
\usage{
CutSpatialNetwork(
  tiledir,
  host = "127.0.0.1",
  port = 8080L,
  container_width = "800px",
  container_height = "800px",
  overwrite_network_json = TRUE,
  verbose = TRUE,
  ...
)

export_graph(object, sampleID = 1L, outdir, verbose = TRUE)
}
\arguments{
\item{tiledir}{A directory containing network data}

\item{host}{A string with a valid IPv4 or IPv6 address to listen on,
Defaults to localhost "127.0.0.1"}

\item{port}{An integer that indicates the port to listen on.}

\item{container_width, container_height}{Set height and width of container}

\item{overwrite_network_json}{Logical specifying if the JSON file
containing the spatial network should be overwritten after completion}

\item{verbose}{Print messages}

\item{object}{A `Seurat` object created with `STUtility2`}

\item{sampleID}{An integer specifying a sample ID to export
spatial network for}

\item{outdir}{Name of a directory to export JSON file to}
}
\description{
Opens an interactive viewer where

Utility function to prepare data for \code{\link{CutSpatialNetwork}}.
The exported JSON file should be exported to the same directory as
the H&E image tiles generated with \code{\link{TileImage}}
}
