% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/disconnect_regions.R
\name{DisconnectRegions}
\alias{DisconnectRegions}
\alias{DisconnectRegions.default}
\title{Disconnect regions defined by the same label}
\usage{
DisconnectRegions(object, ...)

\method{DisconnectRegions}{default}(object, spots, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{spots}{A character vector with spot IDs present 'object'}

\item{verbose}{Print messages}
}
\description{
Disconnect regions defined by the same label
}
\examples{

library(STUtility2)
library(dplyr)
library(ggplot2)
library(patchwork)

galt_spots_file <- "~/STUtility2/repo/STUtility2/inst/extdata/mousecolon/galt_spots.csv"
galt_spots <- read.csv(galt_spots_file) |>
  as_tibble()

# read coordinates
coordfile <- system.file("extdata/mousecolon/spatial",
                         "tissue_positions_list.csv",
                         package = "STUtility2")
coords <- read.csv(coordfile, header = FALSE) |>
  filter(V2 == 1) |>
  select(V1, V6, V5) |>
  setNames(nm = c("barcode", "x", "y")) |>
  bind_cols(sampleID = 1) |>
  as_tibble()

# Select spots
spots <- galt_spots$barcode[galt_spots$selection == "GALT"]
head(spots)

# Find disconnected regions in GALT spots
disconnected_spot_labels <- DisconnectRegions(coords, spots)

# Add information to coords and plot
gg <- coords |>
  mutate(galt = NA, galt_disconnected = NA)
gg$galt[match(spots, gg$barcode)] <- "galt"
gg$galt_disconnected[match(names(disconnected_spot_labels), gg$barcode)] <- disconnected_spot_labels

p1 <- ggplot(gg, aes(x, y, color = galt))
p2 <- ggplot(gg, aes(x, y, color = galt_disconnected))
p <- p1 + p2 &
  geom_point() &
  theme_void()
p

}
