% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_distance.R
\name{.check_angles}
\alias{.check_angles}
\title{Check that selected angles are valid}
\usage{
.check_angles(spatnet, coords, spots, angles, angles_nbreaks, verbose)
}
\arguments{
\item{spatnet}{A list of tibbles containing spatial networks generated with
\code{\link{GetSpatialNetwork}}}

\item{coords}{A tibble with spot coordinates}

\item{spots}{A character vector with spot IDs}

\item{angles}{A numeric vector of length 2 defining the search area}

\item{angles_nbreaks}{Number of intervals to split search area into}

\item{verbose}{Print messages}
}
\description{
Check that selected angles are valid
}
