% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_seurat_images}
\alias{.check_seurat_images}
\title{Check if a Seurat object contains loaded images}
\usage{
.check_seurat_images(object)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
an error message if conditions are not met
}
\description{
Check if a Seurat object contains loaded images
}
