% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_seurat_object}
\alias{.check_seurat_object}
\title{Check if a Seurat object has been processed with STUtility}
\usage{
.check_seurat_object(object)
}
\arguments{
\item{object}{A Seurat object}

\item{message}{An error message}
}
\description{
Check if a Seurat object has been processed with STUtility
}
