% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform.R
\name{.check_transforms}
\alias{.check_transforms}
\title{Check image transformation tibble}
\usage{
.check_transforms(object, nrows = 1)
}
\arguments{
\item{object}{A `tibble` with transformation parameters}

\item{nrows}{Number of expected rows in `object`}
}
\value{
nothing
}
\description{
Check image transformation tibble
}
