% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform.R
\name{.check_transforms_values}
\alias{.check_transforms_values}
\title{Check values in image transformation tibble}
\usage{
.check_transforms_values(object)
}
\arguments{
\item{object}{a `tibble` with transformation parameters}
}
\value{
nothing
}
\description{
Check values in image transformation tibble
}
