% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.crop_dims}
\alias{.crop_dims}
\title{Crop dims}
\usage{
.crop_dims(dims, crop_area, data, coords_columns)
}
\arguments{
\item{dims}{tibble specifying original image dimensions}

\item{crop_area}{a numeric vector of length 4 specifying a rectangle to crop data by}

\item{data}{list of `tibble` objects with spot coordinates}

\item{coords_columns}{character vector specifying column names for spot coordinates}
}
\value{
a tibble with modified image dimensions \code{dims} and list of `tibble` objects
with cropped spot coordinates
}
\description{
Crop dims
}
