% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.crop_images}
\alias{.crop_images}
\title{Crop images}
\usage{
.crop_images(crop_area, images)
}
\arguments{
\item{crop_area}{a numeric vector of length 4 specifying a rectangle to crop data by}

\item{images}{a list of `raster` objects (images)}
}
\value{
a  list of cropped `raster` objects
}
\description{
Crop images
}
