% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_coords_column}
\alias{.get_coords_column}
\title{Get coords_columns}
\usage{
.get_coords_column(image_use, coords_use)
}
\arguments{
\item{image_use}{string specifying image type to use}

\item{coords_use}{string specifying coordinate type to use}
}
\value{
a character vector with column names for spot coordinates
}
\description{
util function to get appropriate column names for spot coordinates
}
