% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{.get_coords_list}
\alias{.get_coords_list}
\title{Get spot coordinates for selected groups and samples}
\usage{
.get_coords_list(object)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
A list of spot coordinates for each sample
}
\description{
Get spot coordinates for selected groups and samples
}
