% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_dims}
\alias{.get_dims}
\title{Get plot dimensions}
\usage{
.get_dims(data, dims)
}
\arguments{
\item{data}{a list of tibbles containing coordinates and feature values}

\item{dims}{either NULL or a tibble from \code{\link{GetSfaffli(se)@image_info}}}
}
\value{
a tibble with plot dimension information
}
\description{
Get plot dimensions
}
