% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_feature_limits}
\alias{.get_feature_limits}
\title{Get feature limits}
\usage{
.get_feature_limits(data, coords_columns, scale = c("shared", "free"))
}
\arguments{
\item{data}{a list of tibbles containing coordinates and feature values}

\item{coords_columns}{character vector specifying the column names of the coordinates}

\item{scale}{one of "free" or "shared". "free" will fix the value limits for each
feature separately and "shared" will fix the value limits for all plots to be identical}
}
\value{
a list of tibbles with feature value limits
}
\description{
Get feature limits
}
