% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_limits}
\alias{.get_limits}
\title{Get limits for pixel coordinates}
\usage{
.get_limits(x)
}
\arguments{
\item{x}{a tibble with spot coordinates}
}
\value{
a tibble with limits
}
\description{
Get limits for pixel coordinates
}
