% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{.get_spots_list}
\alias{.get_spots_list}
\title{Get spot IDs for selected groups and samples}
\usage{
.get_spots_list(object, sel_groups, column_name, split_by_sample = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{sel_groups}{A character vector with selected labels in \code{column_name}}

\item{column_name}{A character vector specifying the name of a column in \code{object}
meta data slot}
}
\value{
A list of lists with spot IDs for selected groups split by sample
}
\description{
Get spot IDs for selected groups and samples
}
