% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.gg_color_hue}
\alias{.gg_color_hue}
\title{Generate colors}
\usage{
.gg_color_hue(n)
}
\arguments{
\item{n}{integer specifying the number of colors to generate}
}
\description{
Generate colors
}
