% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.inject_images}
\alias{.inject_images}
\title{Inject images to list of `ggplot` objects}
\usage{
.inject_images(
  image_use,
  features,
  arrange_features,
  wrapped_plots,
  images,
  ncol,
  return_plot_list,
  blend = FALSE
)
}
\arguments{
\item{image_use}{string specifying image type to use}

\item{features}{a character vector of features names}

\item{arrange_features}{a string specifying how the `patchwork` should be arranged}

\item{wrapped_plots}{list of `ggplot` objects}

\item{images}{list of `raster` images}

\item{ncol}{number of columns to arrange plots by to produce a `patchwork`}

\item{return_plot_list}{logical specifying if the result should be returned
as a `patchwork` or a list of `ggplot` objects}

\item{blend}{logical specifying if blend is activated}
}
\value{
`patchwork` or a list of `ggplot` objects
}
\description{
Inject images to list of `ggplot` objects
}
