% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_regions.R
\name{.is_disconnected}
\alias{.is_disconnected}
\title{Check if spatial network is disconnected}
\usage{
.is_disconnected(spatnet)
}
\arguments{
\item{spatnet}{A list of tibbles with spot IDs for neighboring spots
generate with \code{\link{GetSpatialNetwork}}.}
}
\description{
Check if spatial network is disconnected
}
