% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial_stats.R
\name{.plot_histogram}
\alias{.plot_histogram}
\title{Add a histogram}
\usage{
.plot_histogram(p_stat_base, x, features)
}
\arguments{
\item{p_stat_base}{An empty `ggplot`}

\item{x}{Data to use for plot}

\item{features}{Selected features}
}
\value{
a `ggplot` object
}
\description{
Add a histogram
}
