% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.prep_data_for_plotting}
\alias{.prep_data_for_plotting}
\title{Check input for compatibility}
\usage{
.prep_data_for_plotting(
  object,
  colors,
  label_by,
  scale,
  arrange_features,
  coords_columns
)
}
\arguments{
\item{object}{a tibble with spatial coordinates and feature values}

\item{colors}{a character vector of colors}

\item{scale}{one of "free" or "shared". "free" will fix the value limits for each
feature separately and "shared" will fix the value limits for all plots to be identical}

\item{arrange_features}{one of "row" or "col". "col" will put the features in columns
and samples in rows and "row" will transpose the arrangement}
}
\description{
Check input for compatibility
}
