% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.spatial_feature_plot}
\alias{.spatial_feature_plot}
\title{Plot numeric features in 2D}
\usage{
.spatial_feature_plot(
  gg,
  nm,
  ftr = NULL,
  feature_limits = NULL,
  colors,
  dims,
  all_features = NULL,
  extreme_colors = NULL,
  pt_size = 1,
  pt_alpha = 1,
  pt_stroke = 0,
  cur_label = NULL,
  coords_columns
)
}
\arguments{
\item{gg}{tibble with spatial coordinates and feature values}

\item{nm}{sample ID}

\item{ftr}{feature name}

\item{feature_limits}{list of tibbles containing information about
the feature value range}

\item{dims}{tibble containing information about the dimensions
of the plotting area}

\item{all_features}{a character vector with all features, only used
for blending features}

\item{extreme_colors}{a character vector with the hex colors, only
used for blending features}

\item{pt_size}{point size passed to geom_point}

\item{pt_alpha}{point opacity ranging from 0 to 1 passed to geom_point.
0 = fully transparent, 1 = fully opaque}

\item{pt-stroke}{point stroke width}
}
\description{
Plot numeric features in 2D
}
