% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.split_data_by_label}
\alias{.split_data_by_label}
\title{Split data by label}
\usage{
.split_data_by_label(data, dims, section_number, label, colors)
}
\arguments{
\item{data}{a list of tibbles containing spot coordinates and label columns}

\item{dims}{a tibble with information about plot dimensions to use}

\item{section_number}{an integer specifying the section to keep for splitting}

\item{label}{a string specifying the name of the label column}

\item{colors}{a character vector of colors that should match the number of
groups in the label column}
}
\value{
a list of tibbles holding spot coordinates, a tibble with information
about plot dimensions and a character vector of colors stored together in a list
}
\description{
Takes a list of tibbles holding spot coordinates and a label column,
then selects one of these based on \code{section_number} to use for
restructuring the data. Each group in the label column will get its
own element in the output list of tibbles, where the label column is
set to the group label or "background". This prepares the data for
highlighting 1 group label at the time to use the \code{split_labels}
option in \code{\link{MapLabels}}.
}
