% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.trim_data}
\alias{.trim_data}
\title{Trim numeric features based on predefined cutoffs}
\usage{
.trim_data(data, features, min_cutoff, max_cutoff)
}
\arguments{
\item{data}{a `tibble` containing spatial coordinates and feature values}

\item{features}{a character vector with feature names present in \code{data}}

\item{min_cutoff, max_cutoff}{numeric specifying a cutoff threshold (quantile)}
}
\value{
a `tibble` with trimmed feature values
}
\description{
Trim numeric features based on predefined cutoffs
}
