% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{.validate_column_name}
\alias{.validate_column_name}
\title{validate a selected column name}
\usage{
.validate_column_name(object, column_name)
}
\arguments{
\item{object}{A Seurat object}

\item{column_name}{A character specifying a column name}
}
\value{
nothing
}
\description{
validate a selected column name
}
