% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{.validate_selected_labels}
\alias{.validate_selected_labels}
\title{validate selected labels}
\usage{
.validate_selected_labels(object, sel_groups, column_name)
}
\arguments{
\item{object}{A Seurat object}

\item{sel_groups}{A character vector with selected labels in \code{column_name}}

\item{column_name}{A character vector specifying the name of a column in \code{object}
meta data slot}
}
\value{
A character vector with selected labels in \code{column_name}
}
\description{
validate selected labels
}
