% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_for_viewer.R
\name{ExportDataForViewer}
\alias{ExportDataForViewer}
\title{Export data for FeatureViewer}
\usage{
ExportDataForViewer(
  object,
  sampleIDs = NULL,
  outdir = "./",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A `Seurat` object created with STUtility2}

\item{sampleIDs}{An integer vector specifying}

\item{outdir}{A character vector specifying the path to an existing directory}

\item{overwrite}{Overwrite files if they already exists}

\item{verbose}{Print messages}
}
\description{
Export data for FeatureViewer
}
\examples{
\dontrun{
libary(STUtility2)

se_mbrain <- readRDS(system.file("extdata/mousebrain",
                                 "se_mbrain",
                                 package = "STUtility2"))

# Export files to current working directory
ExportDataForViewer(se_mbrain, outdir = "./")
}

}
