% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_server.R
\name{file_server}
\alias{file_server}
\title{Host a file server}
\usage{
file_server(hostDir, host = "127.0.0.1", port = 8080)
}
\arguments{
\item{hostDir}{A directory to host files from}

\item{host}{A string with a valid IPv4 or IPv6 address to listen on,
Defaults to localhost "127.0.0.1"}

\item{port}{An integer that indicates the port to listen on.}
}
\value{
A \code{Beakr} class object server defining a server instance
}
\description{
Hosts a file server from a specified directory. See
[beakr](https://github.com/MazamaScience/beakr) for more information.
}
\examples{
\donttest{

# Host files in current directory
fs <- file_server(hostDir = "./")

# Stop server
beakr::stopServer(fs)

# Or stop all servers
beakr::stopAllServers()
}

}
