% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial_viewer.R
\name{ftrviewer}
\alias{ftrviewer}
\title{Open a feature viewer react application}
\usage{
ftrviewer(
  host = "127.0.0.1",
  port = "8080",
  sampleID = 1,
  values,
  opacities,
  opacity = 1,
  range,
  scaleByOpacity = FALSE,
  isNumeric = TRUE,
  useLasso = FALSE,
  levels,
  categories = c("cat1", "cat2", "cat3"),
  colors = RColorBrewer::brewer.pal(n = 9, name = "Spectral"),
  container_width = 800,
  container_height = 800,
  elementId = NULL
)
}
\description{
This function will open a react application that listens can
be used to interactively visualize categorical or numeric features
as spatial maps.
}
