% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform.R
\name{generate_rigid_transform}
\alias{generate_rigid_transform}
\title{Utility function to generate a tibble with image transformations}
\usage{
generate_rigid_transform(
  sampleID = 1,
  mirror_x = FALSE,
  mirror_y = FALSE,
  angle = 0,
  tr_x = 0,
  tr_y = 0
)
}
\arguments{
\item{sampleID}{an integer specifying a sample ID}

\item{mirror_x, mirror_y}{logical specifying if the image and spots should be mirrored
along the x- and/or y-axis}

\item{angle}{Numeric value specifying the degree of rotation. Use negative angles
for  counter-clockwise rotation.}
}
\value{
a tibble with image transformations
}
\description{
Utility function to generate a tibble with image transformations
}
