% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{GetCoordinates}
\alias{GetCoordinates}
\alias{GetCoordinates,Staffli-method}
\alias{GetCoordinates,Staffli,Staffli-method}
\alias{GetCoordinates,Seurat-method}
\alias{GetCoordinates,Seurat,Seurat-method}
\title{Method to extract coordinates}
\usage{
GetCoordinates(object)

\S4method{GetCoordinates}{Staffli}(object)

\S4method{GetCoordinates}{Seurat}(object)
}
\arguments{
\item{object}{A \code{Staffli} or \code{Seurat} object}
}
\value{
A \code{tbl} with spot coordinates
}
\description{
Method to extract coordinates
}
\examples{

# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch Staffli object
staffli_object <- GetStaffli(se_mbrain)

# Fetch images from a Staffli object
coordinates <- GetCoordinates(staffli_object)
coordinates


# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch images from a Seurat object
coordinates <- GetCoordinates(se_mbrain)
coordinates

}
