% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{GetImageInfo}
\alias{GetImageInfo}
\alias{GetImageInfo,Staffli-method}
\alias{GetImageInfo,Seurat-method}
\title{Method to extract image info}
\usage{
GetImageInfo(object)

\S4method{GetImageInfo}{Staffli}(object)

\S4method{GetImageInfo}{Seurat}(object)
}
\arguments{
\item{object}{A \code{Staffli} or \code{Seurat} object}
}
\value{
A \code{tbl} with image information
}
\description{
Method to extract image info
}
\examples{

# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch Staffli object
staffli_object <- GetStaffli(se_mbrain)

# Fetch images from a Staffli object
image_info <- GetImageInfo(staffli_object)
image_info


# Fetch images from a Seurat object
image_info <- GetImageInfo(se_mbrain)
image_info

}
