% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{GetImages}
\alias{GetImages}
\alias{GetImages,Staffli-method}
\alias{GetImages,Staffli,Staffli-method}
\alias{GetImages,Seurat-method}
\alias{GetImages,Seurat,Seurat-method}
\title{Method to extract images}
\usage{
GetImages(object, image_use = c("raw", "transformed"))

\S4method{GetImages}{Staffli}(object, image_use = "raw")

\S4method{GetImages}{Seurat}(object, image_use = c("raw", "transformed"))
}
\arguments{
\item{object}{A \code{Staffli} or \code{Seurat} object}

\item{image_use}{A string specifying the image type to get}
}
\value{
A \code{list} of images in \code{raster} format
}
\description{
Method to extract images
}
\examples{

# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch Staffli object
staffli_object <- GetStaffli(se_mbrain)

# Fetch images from a Staffli object
images <- GetImages(staffli_object)


# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch images from a Seurat object
images <- GetImages(se_mbrain)

}
