% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{GetScaleFactors}
\alias{GetScaleFactors}
\alias{GetScaleFactors,Staffli-method}
\alias{GetScaleFactors,Staffli,Staffli-method}
\alias{GetScaleFactors,Seurat-method}
\alias{GetScaleFactors,Seurat,Seurat-method}
\title{Method to extract image scale factors}
\usage{
GetScaleFactors(object)

\S4method{GetScaleFactors}{Staffli}(object)

\S4method{GetScaleFactors}{Seurat}(object)
}
\arguments{
\item{object}{A \code{Staffli} or \code{Seurat} object}
}
\description{
Method to extract image scale factors
}
\examples{

# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch Staffli object
staffli_object <- GetStaffli(se_mbrain)

# Fetch images from a Staffli object
scalefactors <- GetScaleFactors(staffli_object)
scalefactors


# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch images from a Seurat object
scalefactors <- GetScaleFactors(se_mbrain)
scalefactors

}
