% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{GetStaffli}
\alias{GetStaffli}
\alias{GetStaffli,Seurat-method}
\alias{GetStaffli,Seurat,Seurat-method}
\title{Method used to extract a Staffli object from the tools slot of a
Seurat object}
\usage{
GetStaffli(object)

\S4method{GetStaffli}{Seurat}(object)
}
\arguments{
\item{object}{A Seurat object}
}
\description{
Method used to extract a Staffli object from the tools slot of a
Seurat object
}
