% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_images.R
\name{ImagePlot}
\alias{ImagePlot}
\title{Plot H&E images}
\usage{
ImagePlot(
  object,
  label_by = NULL,
  type = "raw",
  section_numbers = NULL,
  ncol = NULL,
  mar = c(1, 1, 1, 1)
)
}
\arguments{
\item{object}{a Seurat object}

\item{label_by}{a string specifying a meta data column to label plots by}

\item{type}{image type, for example 'raw'}

\item{section_numbers}{an integer vector with section numbers to plot}

\item{ncol}{integer value specifying the number of columns in the plot
grid}

\item{mar}{margins around each plot. See \code{\link{par}} for details.}
}
\value{
draws a plot of the H&E images
}
\description{
Plot H&E images
}
\examples{

library(STUtility2)

# Load example Visium data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "STUtility2"))
se_mcolon <- readRDS(system.file("extdata/mousecolon", "se_mcolon", package = "STUtility2"))
se_merged <- MergeSTData(se_mbrain, se_mcolon)

# ImagePlot will throw an error if no images are loaded
\dontrun{
ImagePlot(se_merged)
}

# Load images
se_merged <- LoadImages(se_merged)
ImagePlot(se_merged)

# Plot only selected tissue sections
ImagePlot(se_merged, section_numbers = 1)

# Change order of plot
ImagePlot(se_merged, section_numbers = 2:1)

# Add a sample_id column and label plots
se_merged$sample_id <- ifelse(GetStaffli(se_merged)@meta_data$sampleID == 1, "brain", "colon")
ImagePlot(se_merged, label_by = "sample_id")

}
