% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_image.R
\name{ImageTransform}
\alias{ImageTransform}
\title{Function used to apply rotations and translations to an image object of class `magick-image`}
\usage{
ImageTransform(im, angle, xy_offset)
}
\arguments{
\item{im}{An image of class `magick-image`}

\item{angle}{An integer value specifying the rotation angle (0-360)}

\item{xy_offset}{A numeric vector of length 2 providing the offsets along
the x- and y-axis given as pixels. These values should not exceed the image dimensions.}
}
\value{
An object of class `magick-image`
}
\description{
This function takes an `magick-image` object as input and applies translations
 defined by the \code{angle} and \code{xy_offset} arguments. The output image dimensions
 will remain the same as the input image dimensions, meaning that the transformation might
 result in cropping the image. If you don't want this behavior, you should use `image_rotate`
 from the `magick` R package instead.
}
\examples{

library(magick)
library(STUtility2)
lowresimagefile <- system.file("extdata/mousebrain/spatial",
                               "tissue_lowres_image.png",
                               package = "STUtility2")
im <- image_read(lowresimagefile)

# rotate image 45 degrees clockwise, move image 100 pixels to the right and 100 pixels down
im_transformed <- ImageTransform(im, angle = 45, xy_offset = c(100, 100))
im_transformed

# rotate image 45 degrees counter-clockwise, move image 100 pixels to the left and 20 pixels up
im_transformed <- ImageTransform(im, angle = -45, xy_offset = c(-100, 20))
im_transformed

}
\concept{tranforms}
