% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadAndMergeMatrices}
\alias{LoadAndMergeMatrices}
\title{Create a merged expression matrix from multiple expression matrices stored on disk}
\usage{
LoadAndMergeMatrices(samplefiles, verbose = TRUE)
}
\arguments{
\item{samplefiles}{File paths to}

\item{verbose}{Print messages}
}
\value{
A sparse matrix of class 'dgCMatrix'
}
\description{
The merging makes sure that all genes detected are present in the merged output.
This means that if a gene is missing in a certain dataset, the spots in that
dataset will be assigned with 0 expression.
}
\examples{
\donttest{
samples <- list.files(...)
mergedMatrix <- LoadAndMergeMatrices(samples)
}

}
