% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadAnnotationCSV}
\alias{LoadAnnotationCSV}
\title{Load annotations from a CSV file}
\usage{
LoadAnnotationCSV(paths)
}
\arguments{
\item{paths}{A character vector with paths to CSV files}
}
\value{
A \code{data.frame} object with barcode IDs and annotations
}
\description{
Takes a character vector with paths to CSV files and 
returns a \code{data.frame}. The resulting \code{data.frame}
can be added to a \code{Seurat} object using the \code{AddMetaData}
function.
}
\details{
NB: The files need to be loaded in the correct order. For example, 
if you have an object with 5 tissue sections, you should provide
5 \code{paths} in the correct order.
}
\examples{

library(semla)

se_mcolon <- readRDS(system.file("extdata/mousecolon", "se_mcolon", package = "semla"))
se_merged <- MergeSTData(se_mcolon, y = list(se_mcolon, se_mcolon))

# Get annotation file(s)
annotation_files <- system.file("extdata/mousecolon", 
                                "galt_spots.csv", 
                                package = "semla") |> rep(3)

# Load annotations
annotations <- LoadAnnotationCSV(annotation_files)
head(annotations)

# Edit data.frame if needed
annotations <- annotations |> 
  rename(test = selection)

# Add annotations to meta data
se_merged <- AddMetaData(se_merged, metadata = annotations)

# Plot new categorical variable
MapLabels(se_merged, column_name = "test")

}
\seealso{
Other pre-process: 
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadImages}()},
\code{\link{LoadScaleFactors}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}
}
\concept{pre-process}
