% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadImageData}
\alias{LoadImageData}
\title{Read image data}
\usage{
LoadImageData(images, jsonfiles, verbose = TRUE)
}
\arguments{
\item{images}{An object of class `tibble` containing paths to images in PNG format, with
one row per sample. Paths should specify `.png` files output by spaceranger such as
`tissue_lowres_image.png` or `tissue_hires_image.png`. You do not have to load both
H&E images.}

\item{jsonfiles}{A character vector with file paths. Paths should specify `.json` files containing
scalefactors output by spaceranger.}

\item{verbose}{Print messages}
}
\value{
An object of class `DataFrame``
}
\description{
Load image related data from \strong{'tissue_hires_image.png'} files.
}
\examples{
\donttest{
# Collect spaeranger output files
lowres.imagefiles <- c(system.file("extdata/mousebrain/spatial", "tissue_lowres_image.png", package = "STUtility2"),
             system.file("extdata/mousecolon/spatial", "tissue_lowres_image.png", package = "STUtility2"))
hires.imagefiles <- c(system.file("extdata/mousebrain/spatial", "tissue_hires_image.png", package = "STUtility2"),
             system.file("extdata/mousecolon/spatial", "tissue_hires_image.png", package = "STUtility2"))
jsonfiles <- c(system.file("extdata/mousebrain/spatial", "scalefactors_json.json", package = "STUtility2"),
             system.file("extdata/mousecolon/spatial", "scalefactors_json.json", package = "STUtility2"))

# Load lowres/hires image data for two samples
imgFiles <- tibble::tibble(lowres.imagefiles, hires.imagefiles)
imgData <- LoadImageData(imgFiles, jsonfiles)

# Load lower image data only
imgFiles <- tibble::tibble(lowres.imagefiles)
imgData <- LoadImageData(imgFiles, jsonfiles)
}
}
\seealso{
Other pre-process: 
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadSpatialCoordinates}()}
}
\concept{pre-process}
