% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadImageInfo}
\alias{LoadImageInfo}
\title{Load image information}
\usage{
LoadImageInfo(imgfiles)
}
\arguments{
\item{imgfiles}{A character vector with file paths. Paths should specify PNG or JPEG files containing
images created with spaceranger. Alternatively, the character vector can contain URLs to images.}
}
\value{
A \code{tbl} object with image info
}
\description{
Load image information
}
\examples{
library(semla)

imgs <-
  Sys.glob(paths = paste0(system.file("extdata", package = "semla"),
                          "/*/spatial/tissue_lowres_image.jpg"))
image_info <- LoadImageInfo(imgs)
image_info

}
\seealso{
Other pre-process: 
\code{\link{CreateMultiModalObject}()},
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadAnnotationCSV}()},
\code{\link{LoadImages}()},
\code{\link{LoadScaleFactors}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}
}
\concept{pre-process}
