% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadScaleFactors}
\alias{LoadScaleFactors}
\title{Load scale factors}
\usage{
LoadScaleFactors(scalefactorfiles)
}
\arguments{
\item{scalefactorfiles}{A character vector with file paths. Paths should specify JSON files containing
scale factors created with spaceranger.}
}
\value{
A \code{tbl} with scale factors
}
\description{
Load coordinates from \strong{'scalefactors_json.json'} files and merge them into a \code{tibble}.
}
\examples{
library(semla)

jsonfiles <-
  Sys.glob(paths = paste0(system.file("extdata", package = "semla"),
                          "/*/spatial/scalefactors_json.json"))
scalefactors <- LoadScaleFactors(jsonfiles)
scalefactors

}
\seealso{
Other pre-process: 
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadImages}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}
}
\concept{pre-process}
