% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadSpatialCoordinates}
\alias{LoadSpatialCoordinates}
\title{Load and merge multiple coordinate tables}
\usage{
LoadSpatialCoordinates(
  coordinatefiles,
  remove_spots_outside_tissue = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{coordinatefiles}{Character vector of file paths. Paths should specify \code{.csv} files output by spaceranger}

\item{remove_spots_outside_tissue}{Should spots outside the tissue be removed?}

\item{verbose}{Print messages}
}
\value{
An object of class \code{tbl} containing spot coordinates
}
\description{
Load coordinates from \strong{'tissue_positions_list.csv'} files and merge them into a \code{tibble}.
}
\details{
The merging process makes sure that all genes detected are present in the merged output.
This means that if a gene is missing in a certain dataset, the spots in that dataset will
be assigned with 0 expression.
}
\examples{

library(semla)

# Load and merge coordinates from two samples
coordinatefiles <-
  c(system.file("extdata/mousebrain/spatial",
                "tissue_positions_list.csv",
                package = "semla"),
    system.file("extdata/mousecolon/spatial",
                "tissue_positions_list.csv",
                package = "semla"))
coordinates <- LoadSpatialCoordinates(coordinatefiles)

}
\seealso{
Other pre-process: 
\code{\link{CreateMultiModalObject}()},
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadAnnotationCSV}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadImages}()},
\code{\link{LoadScaleFactors}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}
}
\concept{pre-process}
