% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/local_G.R
\name{RunLocalG}
\alias{RunLocalG}
\alias{RunLocalG.default}
\alias{RunLocalG.Seurat}
\title{Calculate local G statistic}
\usage{
RunLocalG(object, ...)

\method{RunLocalG}{default}(
  object,
  spatnet,
  alternative = c("two.sided", "greater", "less"),
  return_as_tibble = TRUE,
  verbose = TRUE,
  ...
)

\method{RunLocalG}{Seurat}(
  object,
  features,
  alternative = NULL,
  store_in_metadata = TRUE,
  assay_name = "GiScores",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{spatnet}{A list of tibbles containing spatial networks generated with
\code{\link{GetSpatialNetwork}}}

\item{alternative}{A string specifying the alternative hypothesis: "two.sided",
"greater" or "less". By default, only the local G scores are returned. If an
alternative test is specified, the function will return both local G scores and
p-values. Note that p-values are adjusted for multiple hypothesis testing within
each feature using "BH" correction. If you want to adjust p-values with a different
strategy, you can compute the p-values directly from the local G z-scores.}

\item{return_as_tibble}{Logical specifying whether the results should be returned
as an object of class \code{tbl}}

\item{verbose}{Print messages}

\item{features}{A character vector of feature names fetchable with
\code{\link{FetchData}}}

\item{store_in_metadata}{A logical specifying if the results should be
returned in the meta data slot. If set to FALSE, the results will instead
be returned as an assay named by \code{assay_name}. If a statistical test
is applied, the adjusted p-values will be returned to the meta data slot if
\code{store_in_metadata = TRUE} and in the \code{@misc} slot of the assay
if \code{store_in_metadata = FALSE}.}

\item{assay_name}{Name of the assay if \code{store_in_metadata=FALSE}}
}
\value{
An object with local G statistics
}
\description{
This local spatial statistic measures the concentration of
high or low values for a given region. This can for example be used to
define spatial structures in a tissue section based on the values of
selected features. Furthermore, the local G statistic can be used for
high/low clustering of data.

NB: This function only calculates G, not G star.
}
\section{default method}{

Takes a matrix-like object with one feature per column and a list of spatial
networks generated with \code{\link{GetSpatialNetwork}} and computes the local G
scores and optionally p-values. The G scores are prefixed with "Gi" and p-values
are prefixed with one of "Pr(z <!=", "Pr(z " or "Pr(z <" depending on the chosen
test.
}

\section{Seurat}{

Takes a Seurat object as input and returns the local G scores for a
selected number of features and optionally p-values. The G scores are
prefixed with "Gi" and p-values are prefixed with one of "Pr(z <!=",
"Pr(z " or "Pr(z <" depending on the chosen test.
}

\examples{
library(semla)
library(tibble)
library(dplyr)
library(ggplot2)
library(viridis)

# read coordinates
coordfile <- system.file("extdata/mousebrain/spatial",
                         "tissue_positions_list.csv",
                         package = "semla")
coords <- read.csv(coordfile, header = FALSE) |>
  filter(V2 == 1) |>
  select(V1, V6, V5) |>
  setNames(nm = c("barcode", "x", "y")) |>
  bind_cols(sampleID = 1) |>
  as_tibble()

# get spatial network
spatnet <- GetSpatialNetwork(coords)

# Load expression data
feature_matrix <- system.file("extdata/mousebrain",
                              "filtered_feature_bc_matrix.h5",
                              package = "semla") |>
  Seurat::Read10X_h5()
featureMat <- t(as.matrix(feature_matrix[c("Mgp", "Th", "Nrgn"), ]))
featureMat <- featureMat[coords$barcode, ]
head(featureMat)

# Calculate G scores
g_scores <- RunLocalG(log1p(featureMat), spatnet)
head(g_scores)

# Bind results with coords for plotting
gg <- coords |>
  bind_cols(log1p(featureMat)) |>
  left_join(y = g_scores, by = "barcode")

# Plot some results
p1 <- ggplot(gg, aes(x, y, color = Th)) + ggtitle("Expression")
p2 <- ggplot(gg, aes(x, y, color = `Gi[Th]`)) + ggtitle("Local G")
p3 <- ggplot(gg, aes(x, y, color = -log10(`Pr(z != E(Gi[Th]))`))) + ggtitle("-log10(p-value)")

p <- p1 + p2 + p3 &
  geom_point() &
  theme_void() &
  scale_y_reverse() &
  coord_fixed() &
  scale_colour_gradientn(colours = viridis(n = 9))
p


library(semla)
library(dplyr)

# Load Seurat object with mouse bain data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla"))
se_mbrain <- se_mbrain |>
  ScaleData(verbose = FALSE) |>
  FindVariableFeatures(verbose = FALSE) |>
  RunPCA(verbose = FALSE)

# Calculate G scores
se_mbrain <- RunLocalG(se_mbrain, features = c("Th", "Mgp"), alternative = "greater")

# Plot G scores
MapFeatures(se_mbrain, features = "Gi[Th]")

# high/low clustering
se_mbrain$cluster <- se_mbrain[[]] |>
  mutate(cluster = case_when(
    `Pr(z > E(Gi[Th]))` > 0.05 ~ "Not Significant",
    `Pr(z > E(Gi[Th]))` <= 0.05 & `Gi[Th]` > 0 ~ "High"
  )) |> pull(cluster)
MapLabels(se_mbrain, column_name = "cluster")

}
\references{
\itemize{
\item{
Ord, J. K. and Getis, A. 1995 Local spatial autocorrelation statistics:
distributional issues and an application. Geographical Analysis, 27, 286–306
}
\item{
Bivand RS, Wong DWS 2018 Comparing implementations of global and local indicators
of spatial association. TEST, 27(3), 716–748 \href{10.1007/s11749-018-0599-x}{DOI link}
}
}
}
\seealso{
\href{https://sfdep.josiahparry.com/articles/understanding-emerging-hotspots.html}{Emerging Hot Spot Analysis}
\href{https://r-spatial.github.io/spdep/reference/localG.html}{G and G star local spatial statistics}
\href{https://pro.arcgis.com/en/pro-app/latest/tool-reference/spatial-statistics/h-how-high-low-clustering-getis-ord-general-g-spat.htm}{ESRI Getis-ord General G}

Other spatial-methods: 
\code{\link{CorSpatialFeatures}()},
\code{\link{CutSpatialNetwork}()},
\code{\link{DisconnectRegions}()},
\code{\link{GetSpatialNetwork}()},
\code{\link{RadialDistance}()},
\code{\link{RegionNeighbors}()},
\code{\link{RunLabelAssortativityTest}()},
\code{\link{RunNeighborhoodEnrichmentTest}()}
}
\author{
Ludvig Larsson
}
\concept{spatial methods}
\concept{spatial-methods}
