% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rigid_transform_manual.R
\name{RunAlignment}
\alias{RunAlignment}
\alias{RunAlignment.default}
\alias{RunAlignment.Seurat}
\title{Manually apply rigid transformations to images}
\usage{
RunAlignment(object, ...)

\method{RunAlignment}{default}(
  object,
  container_width = "800px",
  container_height = "650px",
  launch_browser = TRUE,
  ...
)

\method{RunAlignment}{Seurat}(
  object,
  image_height = 400,
  container_width = "800px",
  container_height = "650px",
  launch_browser = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{container_width, container_height}{Width and height of paper widget given in pixels.}

\item{launch_browser}{Should the app be opened in the default browser? Default is TRUE.
Set \code{launch_browser = getOption("shiny.launch.browser", interactive())} to use the RStudio 
built-in browser.}

\item{image_height}{An integer used for rescaling images}

\item{verbose}{Print messages}
}
\value{
An object with aligned images
}
\description{
Opens an interactive viewer where images are interactive objects
that can be rotated, moved, scaled and mirrored. It is often difficult
to automate image registration for H&E images, in particular if tissue
sections are damaged, distorted or folded. In such situations, manual
registration can be very useful.
}
\section{default method}{

Takes a list of images prepared with the \code{\link{prep_image}} function
and send them to the interactive application. When the application stops
(after clicking "quit & save") a tibble is returned with information about
the rigid transformations applied to the images.
}

\section{Seurat method}{

Takes a \code{Seurat} object with at least 2 tissue sections and opens an interactive
shiny application where rigid transformations can be applied to the images.
When the application stops (after clicking "quit & save"), the transformations 
are applied to the images using \code{\link{RigidTransformImages}} and a 
\code{Seurat} object is returned. If no transformations are supplied, the function
will return the input \code{Seurat} object unmodified.
}

\examples{

library(semla)
library(magick)

im_mbrain <- system.file("extdata/mousebrain/spatial",
                         "tissue_lowres_image.jpg",
                         package = "semla")

img1 <- prep_image(im_mbrain |>
                     image_read(),
                   height = 256)
img2 <- prep_image(im_mbrain |>
                     image_read() |>
                     image_flip(),
                   height = 256)

# Align images and find transform
transforms <- RunAlignment(object = list(img1, img2))


se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla"))
se_mcolon <- readRDS(system.file("extdata/mousecolon", "se_mcolon", package = "semla"))
se_merged <- MergeSTData(se_mbrain, se_mcolon) |> 
  LoadImages()

# Run alignment application on a Seurat object
se_merged <- RunAlignment(se_merged)

}
\seealso{
Other image transformations: 
\code{\link{MaskImages}()},
\code{\link{RigidTransformImages}()}

Other transforms: 
\code{\link{CoordAndImageTransform}()},
\code{\link{CoordMirror}()},
\code{\link{CoordTransform}()},
\code{\link{ImageTranslate}()},
\code{\link{RigidTransformImages}()}
}
\author{
Ludvig Larsson
}
\concept{image transformations}
\concept{transforms}
