% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial_stats.R
\name{MapFeaturesSummary}
\alias{MapFeaturesSummary}
\title{Map features spatially and add a summary plot next to it}
\usage{
MapFeaturesSummary(
  object,
  features,
  subplot_type = c("box", "violin", "histogram", "density"),
  image_use = NULL,
  coords_use = "raw",
  crop_area = NULL,
  pt_size = 1,
  pt_alpha = 1,
  pt_stroke = 0,
  scale_alpha = FALSE,
  section_number = NULL,
  label_by = NULL,
  ncol = NULL,
  colors = RColorBrewer::brewer.pal(n = 9, name = "Reds"),
  fill_color = NULL,
  scale = c("shared", "free"),
  override_plot_dims = FALSE,
  max_cutoff = NULL,
  min_cutoff = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{a character vector of features to plot. These features need to be
fetchable with \code{link{FetchData}}}

\item{subplot_type}{Select a summary plot to place next to the spatial plot:
\itemize{
   \item{"box" : boxplot}
   \item{"violin" : violin plot}
   \item{"histogram" : histogram}
   \item{"density" : density histogram}
}}

\item{image_use}{string specifying image type to use}

\item{coords_use}{string specifying coordinate type to use}

\item{crop_area}{a numeric vector of length 4 specifying a rectangular area to crop
the plots by. These numbers should be within 0-1. The x-axis is goes from left=0 to
right=1 and the y axis is goes from top=0 to bottom=1. The order of the values are
specified as follows: \code{crop_area = c(left, top, right, bottom)}. The crop area
will be used on all tissue sections and cannot be set for each section individually.
using crop areas of different sizes on different sections can lead to unwanted side
effects as the point sizes will remain constant. In this case it is better to generate
separate plots for different tissue sections.}

\item{pt_size}{numeric value specifying the point size passed to \code{geom_point}}

\item{pt_alpha}{numeric value between 0 and 1 specifying the point opacity passed
to \code{geom_point}. A value of 0 will make the points completely transparent
and a value of 1 will make the points completely opaque.}

\item{pt_stroke}{numeric specifying the point stroke width}

\item{scale_alpha}{logical specifying if the spot colors should be scaled together with
the feature values. This can be useful when you want to highlight regions with higher
feature values while making the background tissue visible.}

\item{section_number}{an integer select a tissue section number to subset data by}

\item{label_by}{character of length 1 providing a column name in \code{object} with
labels that can be used to provide a title for each subplot. This column should have
1 label per tissue section. This can be useful when you need to provide more detailed
information about your tissue sections.}

\item{ncol}{integer value specifying the number of columns in the output patchwork.
This parameter will only have an effect when the number of features provided is 1.
Otherwise, the patchwork will be arranged based on the \code{arrange_features} parameter.}

\item{colors}{a character vector of colors to use for the color scale. The colors should
preferably consist of a set of colors from a scientific color palette designed for sequential
data. Some useful palettes are available in the \code{RColorBrewer}, \code{viridis} and
\code{scico} R packages.}

\item{fill_color}{Fill color for summary plot}

\item{scale}{a character vector of length 1 specifying one of "shared" or "free" which will
determine how the color bars are structured. If scale is set to "shared", the color bars for
feature values will be shared across samples. If scale is set to "free", the color bars
will be independent.}

\item{override_plot_dims}{a logical specifying whether the image dimensions should
be used to define the plot area. Setting \code{override_plot_dims} can be useful
in situations where the tissue section only covers a small fraction of the capture
area, which will create a lot of white space in the plots.}

\item{min_cutoff, max_cutoff}{a numeric value between 0-1 specifying either a lower
(\code{min_cutoff}) or upper (\code{max_cutoff}) limit for the data using \code{\link{quantile}}.
These arguments can be useful to make sure that the color map doesn't get dominated by outliers.}

\item{...}{Arguments passed to other methods}
}
\value{
a `patchwork` object
}
\description{
This function is a wrapped for \code{\link{MapFeatures}} which allows you to
add a boxplot, histogram, violin plot or a density histogram showing the
distribution of the selected feature next to the spatial feature plot.
}
\details{
Note that currently, only 1 feature can be selected
}
\examples{

se_mbrain <- readRDS(system.file("extdata/mousebrain",
                                 "se_mbrain",
                                 package = "STUtility2"))
se_mbrain$sample_id <- "mousebrain"
se_mcolon <- readRDS(system.file("extdata/mousecolon",
                                 "se_mcolon",
                                 package = "STUtility2"))
se_mcolon$sample_id <- "mousecolon"
se <- MergeSTData(se_mbrain, se_mcolon)

# Create Seurat object
se <- se |>
  NormalizeData()

# Add boxplot
MapFeaturesSummary(se, features = "Nrgn", subplot_type = "box")

# Add violin plot
MapFeaturesSummary(se, features = "Nrgn", subplot_type = "violin")

# Add histogram
MapFeaturesSummary(se, features = "Nrgn", subplot_type = "histogram")

# Add density histogram
MapFeaturesSummary(se, features = "Nrgn", subplot_type = "density")

}
\seealso{
Other spatial-visualization: 
\code{\link{AnglePlot}()},
\code{\link{ImagePlot}()},
\code{\link{MapFeatures}()},
\code{\link{MapLabelsSummary}()}
}
\author{
Lovisa Franzen
}
\concept{spatial-visualization}
