% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_and_merge.R
\name{MergeSTData}
\alias{MergeSTData}
\title{Merge 10x Visium data}
\usage{
MergeSTData(
  x,
  y,
  merge_data = TRUE,
  merge_dr = NULL,
  project = "SeuratProject"
)
}
\arguments{
\item{x}{A `Seurat` object}

\item{y}{A `Seurat` object or a list of `Seurat` objects}

\item{merge_data}{Merge the data slots instead of just merging the counts
(which requires re-normalization); this is recommended if the same normalization
approach was applied to all objects. See \code{\link{merge}} for details.}

\item{merge_dr}{Merge specified DimReducs that are present in all objects; will
only merge the embeddings slots for the first N dimensions that are shared across
all objects. See \code{\link{merge}} for details.}

\item{project}{\code{\link{Project}} name for the `Seurat` object}
}
\value{
A merged `Seurat` object
}
\description{
Merges two or more Seurat objects containing SRT data while making sure that the
spatial data (images and spot coordinates) are handled correctly.
}
\details{
NB: If you use the generic \code{\link{merge}} function you will not be able
to use any of the `semla` visualization methods on the output object as
the `Staffli` object will be broken.
}
\examples{
se_mbrain <- readRDS(system.file("extdata",
                                 "mousebrain/se_mbrain",
                                 package = "semla"))
se_mcolon <- readRDS(system.file("extdata",
                                 "mousecolon/se_mcolon",
                                 package = "semla"))

se_mbrain
se_mcolon

# Merge a mousebrain dataset with two mousecolon datasets
se_merged <- MergeSTData(x = se_mbrain, y = se_mcolon)
se_merged

# Plot H&E images
ImagePlot(se_merged |> LoadImages(verbose = FALSE))

}
\seealso{
Other subset-and-merge: 
\code{\link{SubsetSTData}()}
}
\author{
Ludvig Larsson
}
\concept{subset-and-merge}
