% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ModifyPatchworkTitles}
\alias{ModifyPatchworkTitles}
\title{Modify patchwork titles}
\usage{
ModifyPatchworkTitles(p, titles)
}
\arguments{
\item{p}{A \code{patchwork} object}

\item{titles}{A character vector matching the number of patches in \code{p}}
}
\value{
A \code{patchwork} object
}
\description{
Takes a \code{patchwork} object as input produced with \code{\link{MapFeatures}}
or \code{\link{MapLabels}} and moves the legends to the right side.
}
\examples{
library(semla)

se_mbrain <- readRDS(system.file("extdata/mousebrain", 
                                 "se_mbrain", 
                                 package = "semla"))
                                 
# With MapFeatures
p <- MapFeatures(se_mbrain, c("nFeature_Spatial", "nCount_Spatial"))
ModifyPatchworkTitles(p, titles = c("First", "Second"))

# With MapLabels
p <- MapLabels(se_mbrain, c("sample_id"))
ModifyPatchworkTitles(p, titles = "My title")

}
