% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{MyPlot}
\alias{MyPlot}
\title{Function title, e.g. "Scatter plot"}
\usage{
MyPlot(df, colx, coly, col, ...)
}
\arguments{
\item{df}{data.frame}

\item{colx, coly}{column names specifying numeric vectors to plot}

\item{col}{Color of points}

\item{...}{Optional parameters passed to \code{geom_point}}
}
\value{
Decription of what the function returns, e.g. "A scatter plot"
}
\description{
Function description, e.g. "This function takes two numeric vectors
of equal length and returns a scatter plot"
}
\details{
More details about function
}
\section{Section 1}{

Some more details with its own header
}

\section{Section 2}{

Some more details with its own header
}

\examples{
MyPlot(df = mtcars, colx = "cyl", coly = "hp", col = "red")

}
