% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_neighborhood_stats.R
\name{RunNeighborhoodEnrichmentTest}
\alias{RunNeighborhoodEnrichmentTest}
\title{Neighborhood Enrichment Analysis}
\usage{
RunNeighborhoodEnrichmentTest(
  object,
  column_name,
  column_labels = NA,
  n_permutations = 200,
  nCores = parallel::detectCores() - 1,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{column_name}{Column name in metadata corresponding to label ID of the spots.}

\item{column_labels}{Optional. Provide vector of label IDs to subset your analysis by.
Spots of all other labels will be excluded. Must be more than one. Default is all (NA).}

\item{n_permutations}{Integer specifying number of iterations the labels should be randomized
[default: 200]. Recommended to increase the number of permutations to >=200 for more robust
results. A lower number of permutations will result in high standard deviations and thus more
unreliable z-scores.}

\item{nCores}{Number of cores [default: parallel::detectCores() - 1]}

\item{seed}{A seed for reproducibility [default: 123]}

\item{verbose}{Print messages [default: TRUE]}
}
\value{
A tibble with scores for each label pair.
}
\description{
Performs Neighborhood Enrichment Analysis between spot labels, describing whether
spots of two categories lie next to each other spatially more often than expected
by chance.
}
\details{
This analysis calculates the enrichment score, z-score, based on how often spots
of different categorical labels (specified with \code{column_name}) lies
adjacent to each other. The observed number of edges between the labels is
then compared with the results from a set number of permutations (chosen
with \code{n_permutations}), allowing the calculation of a z-score.
The z-score will indicate if a label pair is overrepresented or underrepresented
as compared to what would be expected to see by chance.
The output of this function is a tibble table that for each label pair contains
information about the observed number of edges (`edges`), the mean of the permuted
results (`perm_mean`), the standard deviation of the permuted results (`perm_sd`),
and the z-score (`z_score`).
}
\examples{
library(semla)

# Read data
se <- readRDS(system.file("extdata/mousebrain",
                          "se_mbrain",
                          package = "semla"))

# Generate clusters
se <- se |>
  NormalizeData() |>
  ScaleData() |>
  FindVariableFeatures() |>
  RunPCA() |>
  FindNeighbors(reduction = "pca", dims = 1:30) |>
  FindClusters()

# Run Neigborhood Enrichment Analysis
res <- RunNeighborhoodEnrichmentTest(object = se,
                                     column_name = "seurat_clusters",
                                     n_permutations = 100,
                                     nCores = 2)

res |> arrange(desc(abs(z_score)))


}
\seealso{
Other spatial-methods: 
\code{\link{CorSpatialFeatures}()},
\code{\link{CutSpatialNetwork}()},
\code{\link{DisconnectRegions}()},
\code{\link{GetSpatialNetwork}()},
\code{\link{RadialDistance}()},
\code{\link{RegionNeighbors}()},
\code{\link{RunLabelAssortativityTest}()},
\code{\link{RunLocalG}()}
}
\author{
Lovisa Franzén
}
\concept{spatial-methods}
