% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_unroll.R
\name{osddu-shiny}
\alias{osddu-shiny}
\alias{osdduOutput}
\alias{renderOsddu}
\title{Shiny bindings for osddu}
\usage{
osdduOutput(outputId, width = "100\%", height = "400px")

renderOsddu(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a osddu}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using osddu within Shiny
applications and interactive Rmd documents.
}
