% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_spatial_network.R
\name{osddu}
\alias{osddu}
\title{Create a react app for digital unrolling}
\usage{
osddu(
  sampleID = 1,
  host = "127.0.0.1",
  port = "8080",
  width = NULL,
  height = NULL,
  elementId = NULL,
  quit = FALSE
)
}
\arguments{
\item{sampleID}{A section ID}

\item{host}{A host address}

\item{port}{A valid port}

\item{width, height}{Width and height of container}

\item{elementId}{The element id of the widget}

\item{quit}{A logical specifying is the app should quit}
}
\description{
This function is used to start an interactive widget for
digital unrolling. It requires a static files server to be
hosted in order to find load the necessary files.
}
