% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper-shiny}
\alias{paper-shiny}
\alias{paperOutput}
\alias{renderPaper}
\title{Shiny bindings for paper}
\usage{
paperOutput(outputId, width = "100\%", height = "400px")

renderPaper(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a paper}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
A \code{shiny.tag.list} output or a \code{shiny.render.function} 
function produced to be used in a \code{shiny} app. \code{paperOutput} 
is used in the UI and \code{renderPaper} on the server side.
}
\description{
Output and render functions for using paper within Shiny
applications and interactive Rmd documents.
}
