% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings_plot.R
\name{PlotFeatureLoadings}
\alias{PlotFeatureLoadings}
\title{Plot feature loadings for dimensional reduction data}
\usage{
PlotFeatureLoadings(
  object,
  dims = 1,
  reduction = "pca",
  nfeatures = 30,
  mode = c("dotplot", "barplot", "heatmap"),
  type = c("positive", "negative", "centered"),
  fill = "lightgrey",
  color = "black",
  bar_width = 0.9,
  pt_size = 4,
  pt_stroke = 0.5,
  linetype = "dashed",
  color_by_loadings = FALSE,
  gradient_colors = RColorBrewer::brewer.pal(n = 9, name = "Reds"),
  ncols = NULL
)
}
\arguments{
\item{object}{An object of class `Seurat`}

\item{dims}{An integer vector of dimensions to plot feature loadings for}

\item{reduction}{A character specifying the dimensionality reduction to use}

\item{nfeatures}{Number of features to show}

\item{mode}{Plot mode. One of "barplot", "dotplot" or "heatmap"}

\item{type}{Mode used to select features:
\itemize{
   \item{"positive" : select features with highest loadings}
   \item{"negative" : select features with lowest loadings}
   \item{"centered" : split selection of features to include top and bottom loadings}
}}

\item{fill}{Fill color for barplot and dotplot}

\item{color}{Border color for barplot and dotplot}

\item{bar_width}{With of barplot provided as a proportion}

\item{pt_size}{Size of points in dotplot}

\item{pt_stroke}{Width of border for points in dotplot}

\item{linetype}{Select a linetype to be used for dotplot, e.g. "solid", "longdash" or "blank"}

\item{color_by_loadings}{Should the fill color of barplot or points reflect the feature loadings?}

\item{gradient_colors}{Colors to use for gradient if `color_by_loadings=TRUE`}

\item{ncols}{Number of columns used for final patchwork}
}
\value{
An object of class `patchwork`
}
\description{
This function can be used to visualize the relative contribution of features (e.g. genes)
to dimensionality reduction vectors. The function provides three modes to draw a
bar plot, a dot plot or a heatmap.
}
\section{Select type}{

For centered dimensionality reduction vectors, such as principal components,
it is best to select features with the highest and lowest loadings. For other types
of dimensionality reduction results, it is likely better to only select the features with
the highest loadings, in which case `type="positive"` is the appropriate choice.
}

\section{Select mode}{

barplots or dotplots can be used to get detailed information about the loadings
for individual factors whereas heatmap is useful to summarize the loadings for
multiple factors. The heatmap option will override the `type` options and only
select the top features. With the heatmap mode, the values data will also be scaled
within each dimensionality reduction vectors to range between 0 and 1.
}

\examples{
{
library(STUtility2)

se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "STUtility2"))

# Run PCA
se_mbrain <- se_mbrain |>
                ScaleData() |>
                RunPCA()

# Plot feature loadings for PC_1 as a dotplot
PlotFeatureLoadings(se_mbrain, reduction = "pca", dims = 1,
                    mode = "dotplot", type = "centered")

# Plot feature loadings for PC_1 and PC_2 as barplots
PlotFeatureLoadings(se_mbrain, reduction = "pca", dims = 1:2,
                    mode = "barplot", type = "centered")

# Plot feature loadings for PC_1 and color bars by loading
PlotFeatureLoadings(se_mbrain, reduction = "pca", dims = 1:2,
                    mode = "barplot", type = "centered",
                    color_by_loadings = TRUE,
                    gradient_colors = RColorBrewer::brewer.pal(n = 11, name = "RdBu") |> rev())
}

}
