% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform_manual.R
\name{prep_image}
\alias{prep_image}
\title{Function used to prepare images for paper JS react app}
\usage{
prep_image(input, height = 256)
}
\arguments{
\item{input}{An object of class 'magick-image' or a path
to an image in png or jpeg format}

\item{width}{width of image sent to react app}
}
\value{
a list with an array buffer, the image dimensions and
the length of the array buffer
}
\description{
Function used to prepare images for paper JS react app
}
