% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_analysis.R
\name{RegionNeighbours.default}
\alias{RegionNeighbours.default}
\title{Autodetect region neighbors}
\usage{
RegionNeighbours.default(
  spatnet,
  spots,
  keep.within.id = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{spatnet}{A list of spatial networks generated with \code{\link{GetSpatialNetwork}}}

\item{spots}{A character vector with spot IDs present in `spatnet`}

\item{keep.within.id}{If set to TRUE, all id spots are kept, otherwise only the spots with outside neighbours are kept}

\item{verbose}{Print messages}
}
\description{
This function allows you to automatically identify neighbors of a selected region.
}
\details{
One way of using method this is to find spots surrounding a certain cluster. First, you need to make sure
the identity of the Seurat object is set to the meta.data column that you want to use, so for example
`se <- SetIdent(se, value = "seurat_clusters")` if you want to use the default seurat clusters.
Then you select the label that defined the region of interest using the `id` parameter, so for example
`ìd = "1"` will use cluster 1 as the region. If you set the `keep.idents` parameter to TRUE, the cluster ids
of the neighbouring spots will be kept in the result, otherwise they will be returned as one single goup.
You can also activate the `keep.within.id` parameter to include all spots of the selected region in the output,
otherwise only the spots along the region border will be kept.
}
\examples{
\dontrun{
}

}
