% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{ReplaceImagePaths}
\alias{ReplaceImagePaths}
\alias{ReplaceImagePaths,Staffli-method}
\alias{ReplaceImagePaths,Staffli,Staffli-method}
\alias{ReplaceImagePaths,Seurat-method}
\alias{ReplaceImagePaths,Seurat,Seurat-method}
\title{Replace image paths}
\usage{
ReplaceImagePaths(object, paths)

\S4method{ReplaceImagePaths}{Staffli}(object, paths)

\S4method{ReplaceImagePaths}{Seurat}(object, paths)
}
\arguments{
\item{object}{A \code{Staffli} or \code{Seurat} object}

\item{paths}{A character vector with image paths}
}
\description{
Replace image paths
}
