% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{show,Staffli-method}
\alias{show,Staffli-method}
\alias{show,Staffli,Staffli-method}
\title{Show method for Staffli objects}
\usage{
\S4method{show}{Staffli}(object)
}
\arguments{
\item{object}{object to print pre-selected attributes for}
}
\description{
Show method for Staffli objects
}
