% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{class}
\name{Staffli-class}
\alias{Staffli-class}
\alias{Staffli}
\title{The Staffli Class}
\description{
The Staffli object is designed to hold information about the spatial data generated in a 10x
Visium SRT experiment. This includes paths to images, spot coordinates as well as loaded images
in \code{raster} format and additional information about these images.
}
\section{Slots}{

\describe{
\item{\code{imgs}}{A character vector of paths to the raw HE images}

\item{\code{rasterlists}}{A list of lists containing images in 'raster' format}

\item{\code{meta_data}}{A tibble with spot coordinates and additional meta data}

\item{\code{image_height}}{The height of the scaled images in pixels}

\item{\code{image_info}}{A tibble with information about the raw images}

\item{\code{scalefactors}}{A tibble with information about scalefactors used to convert coordinates
between the original image and down-scaled representations such as "tissue_lowres_image.jpg"}

\item{\code{pixels_per_um}}{Numeric vector specifying the number of pixels in the raw images that
corresponds to 1 micrometer}

\item{\code{version}}{Package version.}
}}

\author{
Ludvig Larsson
}
