% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{[[,Staffli-method}
\alias{[[,Staffli-method}
\alias{`[[`,Staffli,Staffli-method}
\title{Method used to get meta data from a Staffli object}
\usage{
\S4method{[[}{Staffli}(x, i, j, drop = F)
}
\arguments{
\item{x}{object from which to extract element(s).}

\item{i}{row indices specifying elements to extract.}

\item{j}{column indices specifying elements to extract.}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension.
This only works for extracting elements, not for the replacement.}
}
\description{
Method used to get meta data from a Staffli object
}
