% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{Staffli-methods}
\alias{Staffli-methods}
\alias{.DollarNames.Staffli}
\alias{$.Staffli}
\title{Staffli Methods}
\usage{
\method{.DollarNames}{Staffli}(x, pattern = "")

\method{$}{Staffli}(x, i, ...)
}
\arguments{
\item{x}{A \code{\link{Staffli}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }

\item{i}{Name or index of one or more metadata columns}

\item{...}{Arguments passed to other methods}
}
\value{
A selected metadata column \code{i} for object \code{x}
}
\description{
Methods for \code{\link{Staffli}} objects for generics defined in other
packages
}
\section{Functions}{
\itemize{
\item \code{.DollarNames(Staffli)}: Autocompletion for \code{$} access on a
\code{Staffli} object

\item \code{$}: Metadata access for \code{Staffli} objects

}}
\concept{staffli}
