% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{[[<-,Staffli,ANY,ANY,ANY-method}
\alias{[[<-,Staffli,ANY,ANY,ANY-method}
\alias{`[[<-`,Staffli,Staffli-method}
\title{Method used to set meta data in a \code{Staffli} object}
\usage{
\S4method{[[}{Staffli,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{object in which to replace element(s).}

\item{i}{row indices specifying elements to replace.}

\item{j}{column indices specifying elements to replace.}

\item{...}{additional parameters}

\item{value}{Data to add to meta data data.frame.}
}
\description{
Method used to set meta data in a \code{Staffli} object
}
