% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility2-package.R
\docType{package}
\name{STUtility2-package}
\alias{STUtility2}
\alias{STUtility2-package}
\title{STUtility2: Process, analyze and visualize Spatially Resolved Transcriptomcis data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

STUtility2 collects useful tools for Spatially Resolved Transcriptomics data analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ludvigla.github.io/STUtility2/}
}

}
\author{
\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se} (\href{https://orcid.org/0000-0003-4209-2911}{ORCID})

Authors:
\itemize{
  \item Lovisa Franzen \email{lovisa.franzen@scilifelab.se} (\href{https://orcid.org/0000-0003-3755-718X}{ORCID})
}

}
\keyword{internal}
