% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeClean}
\alias{ThemeClean}
\title{Clean theme}
\usage{
ThemeClean()
}
\value{
A \code{theme} object
}
\description{
Removes titles, subtitles and legends from a \code{patchwork} object
}
\examples{
library(semla)

se_mbrain <- readRDS(system.file("extdata/mousebrain", 
                                 "se_mbrain", 
                                 package = "semla"))
                                 
# With MapFeatures
MapFeatures(se_mbrain, c("nFeature_Spatial", "nCount_Spatial")) &
   ThemeClean()

}
