% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ThemeLegendRight}
\alias{ThemeLegendRight}
\title{Move legend to right side}
\usage{
ThemeLegendRight()
}
\value{
A \code{theme} object
}
\description{
Move the legends in a \code{patchwork} object to the right hand side of the plots
}
\examples{
library(semla)

se_mbrain <- readRDS(system.file("extdata/mousebrain", 
                                 "se_mbrain", 
                                 package = "semla"))
                                 
# With MapFeatures
MapFeatures(se_mbrain, c("nFeature_Spatial", "nCount_Spatial")) &
   ThemeLegendRight()

}
