% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{UpdateImageInfo}
\alias{UpdateImageInfo}
\title{Update an \code{image_info} tibble}
\usage{
UpdateImageInfo(image_info, scalefactors)
}
\arguments{
\item{image_info}{A \code{tbl} object with image info}

\item{scalefactors}{A \code{tbl} object with scale factors}
}
\value{
A \code{tbl} object with image info
}
\description{
Takes two \code{tbl} objects as input, one with image information produced with 
\code{\link{LoadImageInfo}} and one with scale factors produced with 
\code{\link{LoadScaleFactors}}. The scale factors are used to calculate the 
width and height of the original H&E image which will be added to the output
\code{tbl}.
}
\seealso{
Other pre-process: 
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadImages}()},
\code{\link{LoadScaleFactors}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()}
}
\concept{pre-process}
