% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_semla.R
\name{UpdateSeuratForSemla}
\alias{UpdateSeuratForSemla}
\title{Update a Seurat object for compatibility with semla}
\usage{
UpdateSeuratForSemla(
  object,
  image_type = c("tissue_lowres", "tissue_hires"),
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{Seurat} with Visium data}

\item{image_type}{One of "tissue_lowres" or "tissue_hires", specifying
what H&E image was loaded into the \code{Seurat} object. Only used for 
"VisiumV1" data.}

\item{verbose}{Print messages}
}
\value{
A \code{Seurat} object compatible with semla
}
\description{
In order to make a \code{Seurat} object with \code{semla}, 
you can use \code{AddSemlaToSeurat} to add the required 
data to the \code{Seurat} object. This assumes that 
the \code{Seurat} object contains one or more "VisiumV1" 
object(s) in the \code{images} slot. Alternatively, you
can convert a \code{Seurat} object with "SlideSeq" data.
}
\section{VisiumV1}{

Note that you need to specify what H&E image was loaded, 
one of "tissue_lowres" or "tissue_hires". If this argument 
is incorrect, the tissue coordinates will be misplaced.

Visit the \href{https://ludvigla.github.io/semla/articles/getting_started.html}{getting started}
tutorial on our package website for an example on how to convert a \code{Seurat} object with
VisiumV1 data.
}

\section{VisiumV2}{

Note that you need to specify what H&E image was loaded, 
one of "tissue_lowres" or "tissue_hires". If this argument 
is incorrect, the tissue coordinates will be misplaced.

The \href{https://ludvigla.github.io/semla/articles/getting_started.html}{getting started} 
tutorial applies also for VisiumV2 assays. If you are working with VisiumHD data however, 
we recommend loading the data using \code{semla}'s own functions, as detailed in the
\href{https://ludvigla.github.io/semla/articles/visiumHD.html}{VisiumHD} tutorial.

Note also that when adapting the object from hexagonal grid VisiumV2 to \code{semla}'s
object structure, array coordinates are not available. Thus, some plotting methods 
will not work.
}

\section{SlideSeq}{

For SlideSeq data, there's no additional H&E image provided. If you convert
a Seurat object containing SlideSeq data, all image related functionality 
of \code{semla} will be inaccessible.
}

\examples{
\dontrun{
library(semla)
library(SeuratData)

# Load example Seurat object with VisiumV1/VisiumV2 data (depends on your Seurat version)
InstallData("stxBrain")
brain <- LoadData("stxBrain", type = "anterior1")

# Make Seurat object compatible with semla
brain_semla <- UpdateSeuratForSemla(brain)

# Load example Seurat object with SlideSeq data
InstallData("ssHippo")
slide_seq <- LoadData("ssHippo")

# Make Seurat object compatible with semla
slide_seq <- UpdateSeuratForSemla(slide_seq)
}

}
