% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semla_to_seurat.R
\name{UpdateSeuratFromSemla}
\alias{UpdateSeuratFromSemla}
\title{Update a Seurat object created with \code{semla} for compatibility with \code{Seurat}'s spatial functions}
\usage{
UpdateSeuratFromSemla(object, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{Seurat} with Visium data, created with \code{semla}}

\item{verbose}{Print messages}
}
\value{
A \code{Seurat} object compatible with with \code{Seurat}'s spatial functions
}
\description{
Update a Seurat object created with \code{semla} for compatibility with \code{Seurat}'s spatial functions
}
\examples{

library(semla)

# Load example Visium data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla"))
se_mcolon <- readRDS(system.file("extdata/mousecolon", "se_mcolon", package = "semla"))
se_merged <- MergeSTData(se_mbrain, se_mcolon)

# Update object
se_merged <- UpdateSeuratFromSemla(se_merged)

# Use Seurat functions
SpatialFeaturePlot(se_merged, features = "Nrgn")

}
