% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_spatial_network.R,
%   R/prepare_for_digital_unroll.R
\name{CutSpatialNetwork}
\alias{CutSpatialNetwork}
\alias{export_graph}
\title{Cut spatial networks in folded tissues}
\usage{
CutSpatialNetwork(
  object,
  datadir = NULL,
  sampleID = 1,
  host = "127.0.0.1",
  port = 8080L,
  container_width = "800px",
  container_height = "800px",
  overwrite_network_json = TRUE,
  verbose = TRUE
)

export_graph(object, sampleID = 1L, outdir, verbose = TRUE)
}
\arguments{
\item{object}{A `Seurat` object created with `STUtility2`}

\item{datadir}{A directory containing network data and image tiles}

\item{sampleID}{An integer specifying a sample ID to export
spatial network for}

\item{host}{A string with a valid IPv4 or IPv6 address to listen on,
Defaults to localhost "127.0.0.1"}

\item{port}{An integer that indicates the port to listen on.}

\item{container_width, container_height}{Set height and width of container}

\item{overwrite_network_json}{Logical specifying if the JSON file
containing the spatial network should be overwritten after completion}

\item{verbose}{Print messages}

\item{outdir}{Name of a directory to export JSON file to}
}
\description{
Opens an interactive viewer where a spatial network is visualized on top
of an H&E image.

Utility function to prepare data for \code{\link{CutSpatialNetwork}}.
The exported JSON file should be exported to the same directory as
the H&E image tiles generated with \code{\link{TileImage}}
}
\details{
Each spot is connected to adjacent neighbors by edges and
the edges can be cut by holding the SHIFT key while moving the cursor
across them. Cut edges can be mended by holding the CTRL key while moving
the cursor across them. The aim is to cut edges between spots located in separate
layers. The output is a `tbl_graph` object representing the spatial network
which can be processed further with \code{\link{AdjustTissueCoordinates}}
to perform "digital unrolling".

A more detailed tutorial can be found on the `STUtility2` website.
}
\seealso{
Other spatial-methods: 
\code{\link{CorSpatialFeatures}()},
\code{\link{DisconnectRegions}()},
\code{\link{GetSpatialNetwork}()},
\code{\link{RadialDistance}()},
\code{\link{RegionNeighbors}()},
\code{\link{RunLocalG}()}
}
\author{
Ludvig Larsson
}
\concept{spatial-methods}
