% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.arrange_plots}
\alias{.arrange_plots}
\title{Arrange list of plots}
\usage{
.arrange_plots(plots, by.col = TRUE)
}
\arguments{
\item{plots}{list of ggplot objects}

\item{by.col}{should the plot be arranged in columns?}
}
\value{
a patchwork object containing a grid of plots
}
\description{
Arrange list of plots
}
