% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.arrange_plots}
\alias{.arrange_plots}
\title{Arrange plots}
\usage{
.arrange_plots(
  wrapped_plots,
  images = NULL,
  features = NULL,
  blend = FALSE,
  arrange_features = "col"
)
}
\arguments{
\item{wrapped_plots}{A list of `ggplot` objects}

\item{images}{A list of `raster` objects}

\item{features}{A character vector of feature names}

\item{blend}{A logical specifying if colors are blended or not}

\item{arrange_features}{A string specifying how features should
be arranged}
}
\description{
Arrange plots
}
