% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_distance.R
\name{.check_angles}
\alias{.check_angles}
\title{Check that selected angles are valid}
\usage{
.check_angles(spatnet, coords, spots, angles, angles_length_out, verbose)
}
\arguments{
\item{coords}{A tibble with spot coordinates}

\item{spots}{A character vector with spot IDs}

\item{angles}{A numeric vector of length 2 defining the search area}

\item{angles_length_out}{Number of intervals to split search area into}
}
\description{
Check that selected angles are valid
}
