% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform.R
\name{.check_transforms}
\alias{.check_transforms}
\title{Check image transformation tibble}
\usage{
.check_transforms(object)
}
\arguments{
\item{object}{a `tibble` with transformation parameters}
}
\value{
a logical telling if transformations other
than default have been set
}
\description{
Check image transformation tibble
}
