% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.color_blender}
\alias{.color_blender}
\title{Blend values}
\usage{
.color_blender(data, features, blend_order, feature_limits)
}
\arguments{
\item{data}{a list of tibbles containing coordinates and feature values}

\item{features}{a character vector of feature names}

\item{blend_order}{an integer vector specifying the order to blend values by.
This vector can only take values 1, 2 or 3 and at least two values needs to
be provided.}

\item{feature_limits}{a list of tibbles containing information about the
plot dimensions}
}
\value{
a list of tibbles similar to input data but in which the feature
columns have been replaced by a color vector with blended colors called
\code{encoded_cols}

a list of tibbles containing coordinates and a color vector
}
\description{
Blend values
}
