% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_dims}
\alias{.get_dims}
\title{Get plot dimensions}
\usage{
.get_dims(dims)
}
\arguments{
\item{dims}{either NULL or a tibble from \code{\link{GetSfaffli(se)@image_info}}}
}
\value{
a tibble with plot dimension information
}
\description{
Get plot dimensions
}
