% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_images}
\alias{.get_images}
\title{Validate and get images for plotting}
\usage{
.get_images(object, image_use, section_number, column_name, split_labels)
}
\arguments{
\item{object}{a Seurat object}

\item{image_use}{string specifying images to use}

\item{section_number}{an integer value specifying a section number to subset
images by}

\item{column_name}{a string specifying a column in the meta data slot of the
Seurat object containing a factor with labels to color spots by}

\item{split_labels}{logical specifying if labels should be split}
}
\description{
Validate and get images for plotting
}
