% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.get_limits}
\alias{.get_limits}
\title{Get limits for pixel coordinates}
\usage{
.get_limits(x, coords_columns)
}
\arguments{
\item{x}{a tibble with spot coordinates}

\item{coords_columns}{a character vector specifying the columns
names for the spot coordinate vectors}
}
\value{
a tibble with limits
}
\description{
Get limits for pixel coordinates
}
