% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial_stats.R
\name{.plot_violin}
\alias{.plot_violin}
\title{Add a violin plot}
\usage{
.plot_violin(p_stat_base, x, features, fill_color)
}
\arguments{
\item{p_stat_base}{An empty `ggplot`}

\item{x}{Data to use for plot}

\item{features}{Selected features}

\item{fill_color}{Fill color for violin plot}
}
\value{
a `ggplot` object
}
\description{
Add a violin plot
}
