% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial.R
\name{.spatial_label_plot}
\alias{.spatial_label_plot}
\title{Plot labels in 2D}
\usage{
.spatial_label_plot(
  gg,
  nm,
  lbl,
  colors,
  dims,
  pt_size = 1,
  pt_alpha = 1,
  pt_stroke = 0,
  coords_columns,
  cur_label,
  drop_na = FALSE
)
}
\arguments{
\item{gg}{tibble with spatial coordinates and a label column}

\item{nm}{sample ID}

\item{lbl}{label column name}

\item{colors}{a character vector of colors IDs}

\item{dims}{tibble containing information about the dimensions
of the plotting area}

\item{pt_size}{point size passed to geom_point}

\item{pt_alpha}{point opacity ranging from 0 to 1 passed to geom_point.
0 = fully transparent, 1 = fully opaque}

\item{pt_stroke}{point stroke width}

\item{coords_columns}{character vector of length 2 specifying names of
columns in which the spatial coordinates are stored}

\item{cur_label}{string with a title}

\item{drop_na}{logical specifying whether NA values should be dropped}
}
\value{
a `ggplot` object with a spatial plot
}
\description{
Plot labels in 2D
}
