% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_digital_unroll.R
\name{export_graph}
\alias{export_graph}
\title{Export a spatial graph to a JSON file}
\usage{
export_graph(object, sampleID = 1L, outdir, verbose = TRUE)
}
\arguments{
\item{object}{A `Seurat` object created with `semla`}

\item{sampleID}{An integer specifying a sample ID to export
spatial network for}

\item{outdir}{Name of a directory to export JSON file to}

\item{verbose}{Print messages}
}
\description{
Utility function to prepare data for \code{\link{CutSpatialNetwork}}.
The exported JSON file should be exported to the same directory as
the H&E image tiles generated with \code{\link{TileImage}}
}
\seealso{
CutSpatialNetwork
}
