% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_for_viewer.R
\name{ExportDataForViewer}
\alias{ExportDataForViewer}
\title{Export data for FeatureViewer}
\usage{
ExportDataForViewer(
  object,
  sampleIDs = NULL,
  outdir,
  nCores = detectCores() - 1,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{Seurat} object created with \code{semla}}

\item{sampleIDs}{An integer vector specifying the sampleIDs for the datasets to export. By default,
all samples are exported.}

\item{outdir}{A character vector specifying the path to an existing directory with permission to read
and write files.}

\item{nCores}{Number of cores to use parallel image tiling}

\item{overwrite}{Overwrite files if they already exists}

\item{verbose}{Print messages}
}
\value{
A path to the directory where the data is saved
}
\description{
This is a utility function used to export data required for \code{\link{FeatureViewer}}.
}
\details{
\code{\link{FeatureViewer}} will automatically attempt to export these files every time it's run.
With \code{ExportDataForViewer}, you only have to export the files once and you can provide the output
data path for \code{\link{FeatureViewer}} to look for the required files in that directory.
}
\examples{

library(semla)

se_mbrain <- readRDS(system.file("extdata/mousebrain",
                                 "se_mbrain",
                                 package = "semla"))
se_mbrain <- LoadImages(se_mbrain)

# Export viewer files to a temporary directory
outpath <- ExportDataForViewer(se_mbrain, outdir = tempdir(), nCores = 1, overwrite = TRUE)
outpath

}
\seealso{
Other feature-viewer-methods: 
\code{\link{FeatureViewer}()}
}
\concept{feature-viewer-methods}
