% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial_viewer.R
\name{FeatureViewer}
\alias{FeatureViewer}
\title{Interactive spatial feature viewer}
\usage{
FeatureViewer(
  object,
  datadir = NULL,
  selected_features = NULL,
  sampleIDs = 1,
  host = "127.0.0.1",
  port = 8080L,
  container_width = 800,
  container_height = 800,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{datadir}{A directory spatial data and image tiles}

\item{selected_features}{A character vector of features to select for viewer}

\item{sampleIDs}{A vector of section IDs to use for the viewer}

\item{host}{A string with a valid IPv4 or IPv6 address to listen on,
Defaults to localhost "127.0.0.1"}

\item{port}{An integer that indicates the port to listen on.}

\item{container_width, container_height}{Set height and width of container}

\item{verbose}{Print messages}

\item{overwrite_network_json}{Logical specifying if the JSON file
containing the spatial network should be overwritten after completion}
}
\description{
A more detailed tutorial can be found on the `STUtility2` website.
}
\seealso{
Other spatial-visualization: 
\code{\link{AnglePlot}()},
\code{\link{ImagePlot}()},
\code{\link{MapFeaturesSummary}()},
\code{\link{MapFeatures}()},
\code{\link{MapLabelsSummary}()}
}
\concept{spatial-visualization}
