% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_spatial_viewer.R
\name{FeatureViewer}
\alias{FeatureViewer}
\title{Interactive spatial feature viewer}
\usage{
FeatureViewer(
  object,
  slot = "data",
  datadir = NULL,
  selected_features = NULL,
  sampleIDs = 1,
  host = "127.0.0.1",
  port = 8080L,
  container_width = 800,
  container_height = 800,
  verbose = TRUE
)
}
\arguments{
\item{object}{A `Seurat` object}

\item{slot}{A slot to use for Assay data}

\item{datadir}{A directory spatial data and image tiles}

\item{selected_features}{A character vector of features to select for viewer}

\item{sampleIDs}{A vector of section IDs to use for the viewer}

\item{host}{A string with a valid IPv4 or IPv6 address to listen on,
Defaults to localhost "127.0.0.1"}

\item{port}{An integer that indicates the port to listen on.}

\item{container_width, container_height}{Set height and width of container}

\item{verbose}{Print messages}
}
\value{
A `Seurat` object
}
\description{
`FeatureViewer` opens up an interactive shiny application where
one can zoom and pan the H&E image while overlaying and color selected features.
It is also possible to add new categorical features or modify existing
categorical features using a lasso tool.
}
\details{
The viewer requires a tiled H&E image along with some additional image data to work.

A more detailed tutorial can be found on the `STUtility2` website. You can also
get more detailed instructions by pressing the help icon in the app.
}
\seealso{
TileImage

Other spatial-visualization-methods: 
\code{\link{AnglePlot}()},
\code{\link{ImagePlot}()},
\code{\link{MapFeaturesSummary}()},
\code{\link{MapFeatures}()},
\code{\link{MapLabelsSummary}()},
\code{\link{MapLabels}()}
}
\author{
Ludvig Larsson
}
\concept{spatial-visualization-methods}
