% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform.R
\name{generate_rigid_transform}
\alias{generate_rigid_transform}
\title{Utility function to generate a tibble with image transformations}
\usage{
generate_rigid_transform(
  sampleID = 1,
  mirror_x = FALSE,
  mirror_y = FALSE,
  angle = 0,
  tr_x = 0,
  tr_y = 0,
  scalefactor = 1
)
}
\arguments{
\item{sampleID}{An integer specifying a sample ID}

\item{mirror_x, mirror_y}{Logical specifying if the image and spots should be mirrored
along the x- and/or y-axis}

\item{angle}{Numeric value specifying the degree of rotation. Use negative angles
for counter-clockwise rotation. The value needs to be in the range (-360, 360)}

\item{tr_x, tr_y}{Integer values for translations along the x- and y-axis}

\item{scalefactor}{A numeric value specifying a scaling factor between (0, 3)}
}
\value{
A tibble with image transformations
}
\description{
Utility function to generate a tibble with image transformations
}
\examples{

# Flip along x axis and rotate 45 degrees
generate_rigid_transform(sampleID = 1, mirror_x = TRUE, angle = 45)

# Move image 100 pixels along x and y axes
generate_rigid_transform(sampleID = 1, tr_x = 100, tr_y = 100)

}
