% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{GetImages}
\alias{GetImages}
\alias{GetImages,Staffli-method}
\alias{GetImages,Staffli,Staffli-method}
\alias{GetImages,Seurat-method}
\alias{GetImages,Seurat,Seurat-method}
\title{Method to extract images from a Staffli object}
\usage{
GetImages(object, type = "raw")

\S4method{GetImages}{Staffli}(object, type = "raw")

\S4method{GetImages}{Seurat}(object, type = "raw")
}
\arguments{
\item{object}{A Staffli or Seurat object}

\item{type}{A string specifying the mage type to get}
}
\description{
Method to extract images from a Staffli object
}
