% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\docType{methods}
\name{ImageInfo}
\alias{ImageInfo}
\alias{ImageInfo,Staffli-method}
\alias{ImageInfo,Staffli,Staffli-method}
\title{Method for extracting the raw H&E image information from a 'Staffli' object}
\usage{
ImageInfo(object)

\S4method{ImageInfo}{Staffli}(object)
}
\arguments{
\item{object}{A Staffli object}
}
\description{
Method for extracting the raw H&E image information from a 'Staffli' object
}
