% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_image.R
\name{ImageTransform}
\alias{ImageTransform}
\title{Apply rotations and translations to an image}
\usage{
ImageTransform(im, angle = 0, xy_offset = c(0, 0), scalefactor = 1)
}
\arguments{
\item{im}{An image of class \code{magick-image}}

\item{angle}{An integer value specifying the rotation angle [-360, 360]}

\item{xy_offset}{A numeric vector of length 2 providing the offsets along
the x- and y-axis given as pixels. These values should not exceed the image dimensions.}

\item{scalefactor}{A numeric value specifying a scaling factor between [0, 3]}
}
\value{
An object of class \code{magick-image}
}
\description{
This function takes an \code{magick-image} object as input and applies translations
defined by the \code{angle} and \code{xy_offset} arguments. The output image dimensions
will remain the same as the input image dimensions, meaning that the transformation might
result in cropping the image. If you don't want this behavior, you should use \code{image_rotate}
from the \code{magick} R package instead.
}
\examples{

library(magick)
library(semla)
lowresimagefile <- system.file("extdata/mousebrain/spatial",
                               "tissue_lowres_image.jpg",
                               package = "semla")
im <- image_read(lowresimagefile)

# rotate image 45 degrees clockwise, move image 100 pixels to the right and 100 pixels down
im_transformed <- ImageTransform(im, angle = 45, xy_offset = c(100, 100))
im_transformed

# rotate image 45 degrees counter-clockwise, move image 100 pixels to the left and 20 pixels up
im_transformed <- ImageTransform(im, angle = -45, xy_offset = c(-100, 20))
im_transformed

# shrink image to helf width/height
im_transformed <- ImageTransform(im, scalefactor = 0.5)
im_transformed

}
\author{
Ludvig Larsson
}
\concept{tranforms}
