% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_image.R
\name{ImageTranslate}
\alias{ImageTranslate}
\title{Apply translations}
\usage{
ImageTranslate(im, xy_offset = NULL)
}
\arguments{
\item{im}{An image of class \code{magick-image}}

\item{xy_offset}{A numeric vector of length 2 providing the offsets along
the x- and y-axis given as pixels. These values should not exceed the image dimensions.}
}
\value{
An object of class \code{magick-image}
}
\description{
This function takes an \code{magick-image} object as input and applies translations
 defined by the \code{xy_offset} argument. The output image dimensions will remain the same
 as the input image dimensions, meaning that the transformation might result in cropping
 the image.
}
\examples{

library(magick)
library(semla)
lowresimagefile <- system.file("extdata/mousebrain/spatial",
                               "tissue_lowres_image.jpg",
                               package = "semla")
im <- image_read(lowresimagefile)

# move image 100 pixels to the right and 100 pixels down
im_transformed <- ImageTranslate(im, xy_offset = c(100, 100))
im_transformed

# move image 100 pixels to the left and 20 pixels up
im_transformed <- ImageTranslate(im, xy_offset = c(-100, 20))
im_transformed

}
\seealso{
Other transforms: 
\code{\link{CoordAndImageTransform}()},
\code{\link{CoordMirror}()},
\code{\link{CoordTransform}()},
\code{\link{RigidTransformImages}()},
\code{\link{RunAlignment}()}
}
\author{
Ludvig Larsson
}
\concept{transforms}
