% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_transformation.R
\name{ImageTranslate}
\alias{ImageTranslate}
\title{Function used to apply translations to an image object of class `magick-image`}
\usage{
ImageTranslate(im, xy_offset)
}
\arguments{
\item{im}{An image of class `magick-image`}

\item{xy_offset}{A numeric vector of length 2 providing the offsets along
the x- and y-axis given as pixels. These values should not exceed the image dimensions.}
}
\value{
An object of class `magick-image`
}
\description{
This function takes an `magick-image` object as input and applies translations
 defined by the \code{xy_offset} argument. The output image dimensions will remain the same
 as the input image dimensions, meaning that the transformation might result in cropping
 the image.
}
\examples{

library(magick)
frink <- image_read("https://jeroen.github.io/images/frink.png")

# move image 100 pixels to the right and 100 pixels down
frink_transformed <- ImageTranslate(frink, xy_offset = c(100, 100))

# move image 100 pixels to the left and 20 pixels up
frink_transformed <- ImageTranslate(frink, xy_offset = c(-100, 20))

}
