% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/load_images.R
\name{LoadImages}
\alias{LoadImages}
\alias{LoadImages.default}
\alias{LoadImages.Seurat}
\title{Load images}
\usage{
LoadImages(image_paths, ...)

\method{LoadImages}{default}(image_paths, image_height = 400)

\method{LoadImages}{Seurat}(object, image_height = 400)
}
\arguments{
\item{image_paths}{A character vector of image files}

\item{...}{Arguments passed to other methods}

\item{object}{An object}
}
\description{
Load images
}
\section{Seurat}{

If a \code{Seurat} object is provided, the images will be loaded as rasters
into the \code{Staffli} object that is located in the \code{tools} slot.
}

