% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/load_images.R
\name{LoadImages}
\alias{LoadImages}
\alias{LoadImages.default}
\alias{LoadImages.Seurat}
\title{Load images}
\usage{
LoadImages(object, ...)

\method{LoadImages}{default}(object, image_height = 400, pad_info = NULL, verbose = TRUE, ...)

\method{LoadImages}{Seurat}(object, image_height = 400, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{image_height}{An integer specifying the height of the down-scaled images}

\item{pad_info}{A tibble with information about how the image should be padded}

\item{verbose}{print messages}
}
\value{
An object with images in \code{raster} format
}
\description{
Load H&E images (or any other custom images) required for visualization methods in \code{semla}.
}
\section{default method}{

If a character vector of image paths are provided, the images will be loaded,
then down-scaled based on \code{image_height} and returned as a list of \code{raster}
objects. Only JPEG and PNG images are supported.
}

\section{Seurat}{

If a Seurat object is provided, the images will be loaded as \code{raster} objects
and stored inside the \code{Staffli} object that is located in the \code{tools}
slot.
}

\examples{

library(semla)

# Get paths for example images
mousebrain_jpg <- system.file("extdata/mousebrain", 
                              "spatial/tissue_lowres_image.jpg", 
                              package = "semla")
mousecolon_jpg <- system.file("extdata/mousecolon", 
                              "spatial/tissue_lowres_image.jpg", 
                              package = "semla")

rasters <- LoadImages(c(mousebrain_jpg, mousecolon_jpg))

# Save graphical parameters
oldpar <- par(no.readonly = TRUE)

# plot images
par(mfrow = c(1, 2), mar = c(0, 0, 0, 0))
for (rst in rasters) {
  plot(rst)
}

# Reset graphical parameters
par(oldpar)


library(semla)

# Load example Visium data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla"))
se_mcolon <- readRDS(system.file("extdata/mousecolon", "se_mcolon", package = "semla"))
se_merged <- MergeSTData(se_mbrain, se_mcolon)

# Load images
se_merged <- LoadImages(se_merged)

}
\seealso{
Other pre-process: 
\code{\link{CreateMultiModalObject}()},
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadAnnotationCSV}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadScaleFactors}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}

Other pre-process: 
\code{\link{CreateMultiModalObject}()},
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadAnnotationCSV}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadScaleFactors}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}
}
\concept{pre-process}
