% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/load_images.R
\name{LoadImages}
\alias{LoadImages}
\alias{LoadImages.default}
\alias{LoadImages.Seurat}
\title{Load images}
\usage{
LoadImages(object, ...)

\method{LoadImages}{default}(object, image_height = 400, pad_info = NULL, verbose = TRUE, ...)

\method{LoadImages}{Seurat}(object, image_height = 400, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{image_height}{An integer specifying the height of the down-scaled images}

\item{pad_info}{A tibble with information about how the image should be padded}

\item{verbose}{print messages}
}
\description{
Load H&E images (or any other custom images) required for visualization methods in \code{semla}.
}
\section{Seurat}{

If a Seurat object is provided, the images will be loaded as `raster` objects
and stored inside the \code{Staffli} object that is located in the \code{tools}
slot.
}

\section{default method}{

If a character vector of image paths are provided, the images will be loaded,
then down-scaled based on \code{image_height} and returned as a list of `raster`
objects. Only JPEG and PNG images are supported.
}

\examples{

#' library(semla)

# Load example Visium data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla"))
se_mcolon <- readRDS(system.file("extdata/mousecolon", "se_mcolon", package = "semla"))
se_merged <- MergeSTData(se_mbrain, se_mcolon)

# Load images
se_merged <- LoadImages(se_merged)

}
