% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadAndMergeMatrices}
\alias{LoadAndMergeMatrices}
\title{Load and merge multiple gene expression matrices}
\usage{
LoadAndMergeMatrices(samplefiles, verbose = TRUE)
}
\arguments{
\item{samplefiles}{Character vector of file/directory paths. Paths should specify \code{.h5} or
\code{.tsv}/\code{.tsv.gz} files. Alternatively, the paths could specify directories including \code{barcodes.tsv},
\code{features.tsv} and \code{matrix.mtx} files.}

\item{verbose}{Print messages}
}
\value{
A sparse matrix of class \code{dgCMatrix} or a list of sparse matrices of class \code{dgCMatrix}
}
\description{
Gene expression matrices should have features in rows and spots in columns.
}
\details{
The merging process makes sure that all genes detected are present in the merged output.
This means that if a gene is missing in a certain dataset, the spots in that dataset will
be assigned with 0 expression.

Spot IDs are renamed to be unique. Usually, the spots are named something similar to:
\code{"ACGCCTGACACGCGCT-1", "TACCGATCCAACACTT-1"}. If a "-N" suffix is missing from
the barcode IDs, it will be added.

Since spot barcodes are shared across datasets, there is a risk that some of the spot IDs
will be duplicated after merging. To avoid this, the suffix (e.g. "-1") is replaced by
a unique suffix for each loaded matrix: "-1", "-2", "-3", ...
}
\section{IF data}{

If the provided h5 files store antibody capture data, \code{LoadAndMergeMatrices} will
return a list of matrices. If multiple samples are loaded, the RNA expression matrices and 
antibody capture matrices will be merged and returned as separate elements of the list.
Note that if one or more samples only have RNA expression data, the function will add empty
values for those samples in the merged antibody capture matrix.
}

\examples{

# Load and merge two gene expression matrices
samples <-
  c(
    system.file(
      "extdata/mousebrain",
      "filtered_feature_bc_matrix.h5",
      package = "semla"
    ),
    system.file(
      "extdata/mousecolon",
      "filtered_feature_bc_matrix.h5",
      package = "semla"
    )
  )
mergedMatrix <- LoadAndMergeMatrices(samples)

}
\seealso{
Other pre-process: 
\code{\link{CreateMultiModalObject}()},
\code{\link{LoadAnnotationCSV}()},
\code{\link{LoadImageInfo}()},
\code{\link{LoadImages}()},
\code{\link{LoadScaleFactors}()},
\code{\link{LoadSpatialCoordinates}()},
\code{\link{ReadVisiumData}()},
\code{\link{UpdateImageInfo}()}
}
\concept{pre-process}
