% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadSpatialCoordinates}
\alias{LoadSpatialCoordinates}
\title{Load and merge multiple coordinate tables}
\usage{
LoadSpatialCoordinates(coordinatefiles, verbose = TRUE)
}
\arguments{
\item{coordinatefiles}{Character vector of file paths. Paths should specify `.csv` files output by spaceranger.}

\item{verbose}{Print messages}
}
\value{
An object of class `tbl` (`tibble`)
}
\description{
Load coordinates from \strong{'tissue_positions_list.csv'} files and merge them into a `tibble`.
}
\details{
The merging process makes sure that all genes detected are present in the merged output.
This means that if a gene is missing in a certain dataset, the spots in that dataset will
be assigned with 0 expression.
}
\examples{
\donttest{
# Load and merge coordinates from two samples
coordinatefiles <- c(system.file("extdata/mousebrain/spatial", "tissue_positions_list.csv", package = "STUtility2"),
             system.file("extdata/mousecolon/spatial", "tissue_positions_list.csv", package = "STUtility2"))
coordinates <- LoadSpatialCoordinates(coordinatefiles)
}

}
\seealso{
Other pre-process: 
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadImageData}()}
}
\concept{pre-process}
