% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rigid_transform_manual.R
\name{RunAlignment}
\alias{RunAlignment}
\alias{RunAlignment.default}
\title{Manually apply rigid transformations to images}
\usage{
RunAlignment(object, ...)

\method{RunAlignment}{default}(object, width = "800px", height = "650px", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Opens an interactive viewer where images are interactive objects
that can be rotated, moved, scaled and mirrored. It is often difficult
to automate image registration for H&E images, in particular if tissue
sections are damaged, distorted or folded. In such situations, manual
registration can be very useful.
}
\section{default method}{

Takes a list of images prepared with the \code{\link{prep_image}} function
and send them to the interactive application. When the application stops
(after clicking "quit & save") a tibble is returned with information about
the rigid transformations applied to the images.
}

\examples{
\dontrun{

library(STUtility2)
library(magick)

im_mbrain <- system.file("extdata/mousebrain/spatial",
                         "tissue_hires_image.png",
                         package = "STUtility2")

img1 <- prep_image(im_mbrain |>
                     image_read(),
                   width = 512)
img2 <- prep_image(im_mbrain |>
                     image_read() |>
                     image_flip(),
                   width = 512)

transforms <- RunAlignment(object = list(img1, img2))

}

}
\seealso{
Other image transformations: 
\code{\link{MaskImages}()},
\code{\link{RigidTransformImages}()}
}
\concept{image transformations}
