% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rigid_transform_manual.R
\name{RunAlignment}
\alias{RunAlignment}
\alias{RunAlignment.default}
\title{Manually apply rigid transformations to images}
\usage{
RunAlignment(object, ...)

\method{RunAlignment}{default}(
  object,
  container_width = "800px",
  container_height = "650px",
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{container_width, container_height}{Width and height of paper widget given in pixels.}
}
\description{
Opens an interactive viewer where images are interactive objects
that can be rotated, moved, scaled and mirrored. It is often difficult
to automate image registration for H&E images, in particular if tissue
sections are damaged, distorted or folded. In such situations, manual
registration can be very useful.
}
\section{default method}{

Takes a list of images prepared with the \code{\link{prep_image}} function
and send them to the interactive application. When the application stops
(after clicking "quit & save") a tibble is returned with information about
the rigid transformations applied to the images.
}

\examples{
\dontrun{

library(semla)
library(magick)

im_mbrain <- system.file("extdata/mousebrain/spatial",
                         "tissue_hires_image.jpg",
                         package = "semla")

img1 <- prep_image(im_mbrain |>
                   image_read(),
                   height = 512)
img2 <- prep_image(im_mbrain |>
                   image_read() |>
                   image_flip(),
                   height = 512)

transforms <- RunAlignment(object = list(img1, img2))

}

}
\seealso{
Other image transformations: 
\code{\link{MaskImages}()},
\code{\link{RigidTransformImages}()}

Other transforms: 
\code{\link{CoordAndImageTransform}()},
\code{\link{CoordMirror}()},
\code{\link{CoordTransform}()},
\code{\link{ImageTranslate}()},
\code{\link{RigidTransformImages}()}
}
\author{
Ludvig Larsson
}
\concept{image transformations}
\concept{transforms}
