% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{mcolon_dataset}
\alias{mcolon_dataset}
\title{Visium mouse colon dataset}
\description{
A Visium dataset obtained from a "swiss roll" of a mouse colon
\itemize{
   \item{\strong{mousecolon/filtered_feature_bc_matrix.h5}:\cr gene expression matrix in hdf5 format filtered to include spots under the tissue}
   \item{\strong{mousecolon/spatial/tissue_lowres_image.png}: \cr H&E image (600x541) pixels}
   \item{\strong{mousecolon/spatial/tissue_hires_image.png}: \cr H&E image (2000x1804) pixels}
   \item{\strong{mousecolon/spatial/tissue_positions_list.csv}: \cr CSV file with spot coordinates}
   \item{\strong{mousecolon/spatial/scalefactors_json.json}: \cr JSON file with scalefactors}
   \item{\strong{mousecolon/se_mcolon}: \cr Seurat object stored as an .Rds file with the top 100 most variable features}
}
}
\seealso{
Other datasets: 
\code{\link{mbrain_dataset}}
}
\concept{datasets}
