% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform_manual.R
\name{paper}
\alias{paper}
\title{Create a react app for paper JS in R}
\usage{
paper(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A list of images prepared with \code{\link{prep_image}}}

\item{width}{Width of component}

\item{height}{height of component}

\item{elementId}{Component element ID}
}
\description{
Provided a list of images, this function is used to open an interactive app
built with react and paper JS.
}
