% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper}
\alias{paper}
\title{Create a react app for paper JS in R}
\usage{
paper(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A list of images prepared with \code{\link{prep_image}}}

\item{width}{Width of component}

\item{height}{height of component}

\item{elementId}{Component element ID}
}
\description{
Provided a list of images, this function is used to open an interactive app
built with react and paper JS. The application is mean to be used in a shiny 
application and is used for the \code{\link{RunAlignment}} function provided 
in \code{semla}.
}
\examples{

\dontrun{
library(semla)
library(magick)

im_mbrain <- system.file("extdata/mousebrain/spatial",
                         "tissue_hires_image.jpg",
                         package = "semla")

img1 <- prep_image(im_mbrain |>
                     image_read(),
                   height = 512)
img2 <- prep_image(im_mbrain |>
                     image_read() |>
                     image_flip(),
                   height = 512)

paper(data = list(img1, img2), width = 600, height = 600)
}

}
