% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{plot,Staffli,missing-method}
\alias{plot,Staffli,missing-method}
\title{Plot method for \code{Staffli} objects}
\usage{
\S4method{plot}{Staffli,missing}(
  x,
  image_use = NULL,
  coords_use = c("raw", "transformed"),
  ncol = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{Staffli} object}

\item{image_use}{A string specifying the image to plot}

\item{coords_use}{A character vector of length 2 specifying the coordinates to use}

\item{ncol}{Integer specifying the number of columns in the plot grid}

\item{...}{Additional parameters passed to \code{geom_point}}
}
\value{
No return value, plots the content of a \code{Staffli} object
}
\description{
Plot method for \code{Staffli} objects
}
