% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigid_transform_manual.R
\name{prep_image}
\alias{prep_image}
\title{Function used to prepare images for paper JS react app}
\usage{
prep_image(input, width = 256)
}
\arguments{
\item{width}{width of image sent to react app}

\item{path}{path to image in png or jpeg format}
}
\value{
a list with an array buffer, the image dimensions and
the length of the array buffer
}
\description{
Function used to prepare images for paper JS react app
}
