% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{ReadVisiumData}
\alias{ReadVisiumData}
\title{Read spaceranger output files}
\usage{
ReadVisiumData(infoTable, assay = "Spatial", verbose = TRUE, ...)
}
\arguments{
\item{infoTable}{A `data.frame` or `tbl` with paths to spaceranger output files}

\item{assay}{Assay name (default = "Spatial")}

\item{verbose}{Print messages}

\item{...}{Parameters passed to \code{\link{CreateSeuratObject}}}
}
\description{
This function serves as a wrapper for \code{\link{LoadAndMergeMatrices}} and
\code{\link{LoadSpatialCoordinates}} to load spaceranger output files and create
a Seurat object. The spatial information, i.e. images and spot coordinates, are
stored inside the tools slot of the `Seurat` object in an object called `Staffli`.
}
\examples{
# Assemble spaceranger output files
samples <- Sys.glob(paths = paste0(system.file("extdata", package = "STUtility2"), "/*/filtered_feature_bc_matrix.h5"))
imgs <- Sys.glob(paths = paste0(system.file("extdata", package = "STUtility2"), "/*/spatial/tissue_hires_image.png"))
spotfiles <- Sys.glob(paths = paste0(system.file("extdata", package = "STUtility2"), "/*/spatial/tissue_positions_list.csv"))
json <- Sys.glob(paths = paste0(system.file("extdata", package = "STUtility2"), "/*/spatial/scalefactors_json.json"))

# Create a tibble/data.frame with file paths
library(tibble)
infoTable <- tibble(samples, imgs, spotfiles, json, sample_id = c("mousebrain", "mousecolon"))

# Create Seurat object
se <- ReadVisiumData(infoTable = infoTable)
se

}
\seealso{
Other pre-process: 
\code{\link{LoadAndMergeMatrices}()},
\code{\link{LoadImageData}()},
\code{\link{LoadSpatialCoordinates}()}
}
\concept{pre-process}
