% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semla-package.R
\docType{package}
\name{semla-package}
\alias{semla}
\alias{semla-package}
\title{semla: Tools for Spatially Resolved Transcriptomics Data Processing, Analysis and Visualization}
\description{
A toolkit for Spatially Resolved Transcriptomics data exploration and analysis. semla provides interactive tools for image alignment and data exploration together with several analysis and visualization methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ludvigla.github.io/semla/}
}

}
\author{
\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se} (\href{https://orcid.org/0000-0003-4209-2911}{ORCID})

Authors:
\itemize{
  \item Lovisa Franzen \email{lovisa.franzen@scilifelab.se} (\href{https://orcid.org/0000-0003-3755-718X}{ORCID})
}

Other contributors:
\itemize{
  \item Spatial Research [funder]
}

}
\keyword{internal}
