% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semla-package.R
\docType{package}
\name{semla-package}
\alias{semla}
\alias{semla-package}
\title{semla: Process, analyze and visualize Spatially Resolved Transcriptomcis data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

semla collects useful tools for Spatially Resolved Transcriptomics data analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ludvigla.github.io/semla/}
}

}
\author{
\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se} (\href{https://orcid.org/0000-0003-4209-2911}{ORCID})

Authors:
\itemize{
  \item Lovisa Franzen \email{lovisa.franzen@scilifelab.se} (\href{https://orcid.org/0000-0003-3755-718X}{ORCID})
}

Other contributors:
\itemize{
  \item Spatial Research [funder]
}

}
\keyword{internal}
