% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{show,Staffli-method}
\alias{show,Staffli-method}
\alias{show,Staffli,Staffli-method}
\title{Show method for \code{Staffli} objects}
\usage{
\S4method{show}{Staffli}(object)
}
\arguments{
\item{object}{object to print pre-selected attributes for}
}
\value{
No return value, plots the spot coordinates and alternatively also
the images found in a \code{Staffli} object
}
\description{
Show method for \code{Staffli} objects
}
