% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffli_object.R
\name{[[,Staffli,ANY,ANY-method}
\alias{[[,Staffli,ANY,ANY-method}
\alias{`[[`,Staffli,Staffli-method}
\title{Method used to get meta data from a \code{Staffli} object}
\usage{
\S4method{[[}{Staffli,ANY,ANY}(x, i, j, drop = F)
}
\arguments{
\item{x}{object from which to extract element(s).}

\item{i}{row indices specifying elements to extract.}

\item{j}{column indices specifying elements to extract.}

\item{drop}{If TRUE the result is coerced to the lowest possible dimension.
This only works for extracting elements, not for the replacement.}
}
\description{
Method used to get meta data from a \code{Staffli} object
}
\examples{

# Load example data
se_mbrain <- readRDS(system.file("extdata/mousebrain", "se_mbrain", package = "semla")) |> 
   LoadImages()

# Fetch Staffli object
staffli_object <- GetStaffli(se_mbrain)

# Get meta_data from a staffli object
meta_data <- staffli_object[[]]

}
