% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STUtility2-package.R
\docType{package}
\name{STUtility2-package}
\alias{STUtility2}
\alias{STUtility2-package}
\title{STUtility2: Process, analyze and visualize Spatially Resolved Transcriptomcis data.}
\description{
STUtility2 collects useful tools for Spatially Resolved Transcriptomics data analysis within R.
}
\author{
\strong{Maintainer}: Ludvig Larsson \email{ludvig.larsson@scilifelab.se}

Authors:
\itemize{
  \item Lovisa Franzen \email{lovisa.franzen@scilifelab.se}
}

}
\keyword{internal}
