% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_images.R
\name{TileImage}
\alias{TileImage}
\title{Tile an H&E image}
\usage{
TileImage(
  im,
  outpath = NULL,
  maxZoomLevel = 4,
  maxImgWidth = 10000,
  nCores = detectCores() - 1
)
}
\arguments{
\item{im}{An image of class `magick-image`}

\item{outpath}{A string specifying an output directory to save the tiled image in.
If this is not provided, a temporary directory will be created}

\item{maxZoomLevel}{Max zoom level}

\item{maxImgWidth}{Safety threshold to make sure that the zoom level doesn't get too deep.}

\item{nCores}{Number of cores to use for threading}
}
\value{
Path to tiles
}
\description{
This function takes an image of class `magick-image` and create a tile
map. The size of each tile is 256x256 pixels and the number of zoom levels
are determined rom the
}
\examples{
\donttest{
if (!requireNamespace("shiny"))
  install.packages("shiny")
if (!requireNamespace("leaflet"))
  install.packages("leaflet")
library(magick)
library(shiny)
library(leaflet)

# Load H&E image with magick
imfile <-
  system.file("extdata/mousecolon/spatial",
              "tissue_hires_image.png",
              package = "STUtility2")
im <- image_read(imfile)

# tile image and return path to tiles
tile_res <- TileImage(im, nCores = 2)

# Create a simple viewer with leaflet
ui <- fluidPage(
  leafletOutput("map", height = 512, width = 512),
)

server <- function(input, output, session) {
  addResourcePath("mytiles", "/Users/ludviglarsson/Downloads/tiles")
  output$map <- renderLeaflet({
    leaflet(options = leafletOptions(preferCanvas = TRUE)) \%>\%
      addTiles(urlTemplate = "/mytiles/{z}/{x}_{y}.jpg",
               options = tileOptions(continuousWorld = TRUE,
                                     tileSize = "256",
                                     minZoom = paste0(res$minZoomLevel),
                                     maxZoom = paste0(res$maxZoomLevel)))
  })
}

shinyApp(ui, server)
}

}
