% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stutilityv1_object.R
\name{UpdateSTUtilityV1Object}
\alias{UpdateSTUtilityV1Object}
\title{Update an STUtility v1 object to work with \code{semla}}
\usage{
UpdateSTUtilityV1Object(object, verbose = TRUE)
}
\arguments{
\item{object}{A \code{Seurat} object created with \code{STUtility} v1}

\item{verbose}{Print messages}
}
\value{
A \code{Seurat} object compatible with \code{semla}
}
\description{
Some features will no longer be available when updating an \code{STUtility} v1 object.
\code{STUtility} v1 objects do not store information about scaling factors to convert
between pixels and real distances which will set limitations on some of the
visualization methods and spatial functions. To mitigate these issues, you can
either reload the data from the raw space ranger output files or manually add
the missing information stored in the "scalefactors_json.json" files.
}
\details{
Note that valid image paths need to be available from the \code{STUtility} v1 object.
If not, you can set the paths manually before updating the object:
\code{old_se@tools$Staffli@imgs <- c("path/to/im1.jpg", "path/to/im2.jpg", ...)}
}
\examples{

\dontrun{
se_old_url <- file.path("https://data.mendeley.com/public-files/datasets/kj3ntnt6vb/files",
                        "8f9d3dee-a026-40f0-be66-2df693f9db66/file_downloaded")

# Load old STUtility object
se_old <- readRDS(url(se_old_url))

# Update imgs slot of Staffli object to a valid path or URL (H&E image)
he_img <- file.path("https://data.mendeley.com/public-files/datasets/kj3ntnt6vb/files",
                    "d97fb9ce-eb7d-4c1f-98e0-c17582024a40/file_downloaded")
se_old@tools$Staffli@imgs <- he_img

# Update old STUtility object
se_new <- UpdateSTUtilityV1Object(se_old)
}

}
\author{
Ludvig Larsson
}
